refactor(vm): split builtin implementations by domain
