+++ internal/vm/vm.go
@@ -1,41 +1,22 @@
 package vm
 
 import (
-	"bufio"
-	"crypto/aes"
-	"crypto/cipher"
-	"crypto/pbkdf2"
-	"crypto/rand"
-	"crypto/sha256"
 	"database/sql"
-	"encoding/base64"
-	"encoding/hex"
-	"encoding/json"
 	"fmt"
-	"io"
 	"net"
 	"noxy-vm/internal/ast"
 	"noxy-vm/internal/chunk"
 	"noxy-vm/internal/compiler"
 	"noxy-vm/internal/lexer"
 	"noxy-vm/internal/parser"
-	"noxy-vm/internal/plugin"
 	"noxy-vm/internal/stdlib"
 	"noxy-vm/internal/value"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"reflect"
-	"regexp"
-	"runtime"
-	"runtime/debug"
-	"strconv"
 	"strings"
 	"sync"
-	"time"
 	"unicode/utf8"
-
-	_ "modernc.org/sqlite"
 )
 
 const StackMax = 2048
@@ -144,3280 +125,8 @@ func NewWithShared(shared *SharedState, cfg VMConfig) *VM {
 		netBufferedConns: make(map[int]net.Conn),
 	}
 
-	// Define 'print' native
-	vm.DefineNative("print", func(args []value.Value) value.Value {
-		var parts []string
-		for _, arg := range args {
-			parts = append(parts, arg.String())
-		}
-		fmt.Println(strings.Join(parts, " "))
-		return value.NewNull()
-	})
-
-	// Define 'iprint' native (inline print)
-	vm.DefineNative("iprint", func(args []value.Value) value.Value {
-		var parts []string
-		for _, arg := range args {
-			parts = append(parts, arg.String())
-		}
-		fmt.Print(strings.Join(parts, " "))
-		return value.NewNull()
-	})
-
-	// Concurrency Primitives
-	vm.DefineNative("spawn", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewNull()
-		}
-		fnVal := args[0]
-		if fnVal.Type != value.VAL_FUNCTION {
-			// Only script functions are supported in spawn.
-			fmt.Println("Runtime Error: spawn expects a function")
-			return value.NewNull()
-		}
-
-		threadArgs := args[1:]
-
-		// Create new VM thread sharing state
-		threadVM := NewWithShared(vm.shared, vm.Config)
-
-		// Setup execution
-		var closure *value.ObjClosure
-		if cl, ok := fnVal.Obj.(*value.ObjClosure); ok {
-			closure = cl
-		} else if fn, ok := fnVal.Obj.(*value.ObjFunction); ok {
-			closure = &value.ObjClosure{Function: fn, Upvalues: []*value.ObjUpvalue{}}
-		} else {
-			fmt.Println("Runtime Error: spawn expects a function or closure")
-			return value.NewNull()
-		}
-
-		fnObj := closure.Function
-
-		// Check arity
-		if len(threadArgs) != fnObj.Arity {
-			fmt.Printf("Runtime Error: spawn expected %d args, got %d\n", fnObj.Arity, len(threadArgs))
-			return value.NewNull()
-		}
-
-		// Push Function (Stack slot 0)
-		threadVM.push(fnVal)
-
-		// Push Args
-		for _, arg := range threadArgs {
-			threadVM.push(arg)
-		}
-
-		// Create Frame
-		frame := &CallFrame{
-			Closure: closure,
-			IP:      0,
-			Slots:   0,
-			Globals: nil,
-		}
-
-		// Inherit globals from the function/closure.
-		frame.Globals = fnObj.Globals
-
-		threadVM.frames[0] = frame
-		threadVM.frameCount = 1
-		threadVM.currentFrame = frame
-
-		// Launch Goroutine
-		go func() {
-			defer func() {
-				if r := recover(); r != nil {
-					fmt.Printf("Thread Panic: %v\n%s", r, debug.Stack())
-				}
-			}()
-			err := threadVM.run(1) // Run until finished (frame 0 popped)
-			if err != nil {
-				fmt.Printf("Thread Error: %v\n", err)
-			}
-		}()
-
-		return value.NewNull()
-	})
-
-	vm.DefineNative("make_chan", func(args []value.Value) value.Value {
-		size := 0
-		if len(args) > 0 {
-			if args[0].Type == value.VAL_INT {
-				size = int(args[0].AsInt)
-			}
-		}
-		return value.NewChannel(size)
-	})
-
-	vm.DefineNative("chan_send", func(args []value.Value) value.Value {
-		if len(args) != 2 {
-			return value.NewNull()
-		}
-		if args[0].Type != value.VAL_CHANNEL {
-			return value.NewNull()
-		}
-		ch := args[0].Obj.(*value.ObjChannel).Chan
-		ch <- args[1]
-		return args[1]
-	})
-
-	vm.DefineNative("chan_close", func(args []value.Value) value.Value {
-		if len(args) != 1 {
-			return value.NewNull()
-		}
-		if args[0].Type != value.VAL_CHANNEL {
-			return value.NewNull()
-		}
-		chObj := args[0].Obj.(*value.ObjChannel)
-
-		chObj.Lock.Lock()
-		defer chObj.Lock.Unlock()
-
-		if !chObj.Closed {
-			close(chObj.Chan)
-			chObj.Closed = true
-		}
-		return value.NewNull()
-	})
-
-	vm.DefineNative("chan_is_closed", func(args []value.Value) value.Value {
-		if len(args) != 1 {
-			return value.NewBool(false)
-		}
-		if args[0].Type != value.VAL_CHANNEL {
-			return value.NewBool(false)
-		}
-		chObj := args[0].Obj.(*value.ObjChannel)
-
-		chObj.Lock.Lock()
-		defer chObj.Lock.Unlock()
-
-		return value.NewBool(chObj.Closed)
-	})
-
-	vm.DefineNative("chan_recv", func(args []value.Value) value.Value {
-		if len(args) != 1 {
-			return value.NewNull()
-		}
-		if args[0].Type != value.VAL_CHANNEL {
-			return value.NewNull()
-		}
-		ch := args[0].Obj.(*value.ObjChannel).Chan
-		val, ok := <-ch
-		if !ok {
-			return value.NewNull()
-		}
-		return val
-	})
-
-	vm.DefineNative("make_wg", func(args []value.Value) value.Value {
-		return value.NewWaitGroup()
-	})
-
-	vm.DefineNative("wg_add", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewNull()
-		}
-		if args[0].Type != value.VAL_WAITGROUP {
-			return value.NewNull()
-		}
-		delta := int(0)
-		if args[1].Type == value.VAL_INT {
-			delta = int(args[1].AsInt)
-		}
-		if delta == 0 {
-			return value.NewNull()
-		}
-		wg := args[0].Obj.(*value.ObjWaitGroup).Wg
-		wg.Add(delta)
-		return value.NewNull()
-	})
-
-	vm.DefineNative("wg_done", func(args []value.Value) value.Value {
-		if len(args) != 1 {
-			return value.NewNull()
-		}
-		if args[0].Type != value.VAL_WAITGROUP {
-			return value.NewNull()
-		}
-		wg := args[0].Obj.(*value.ObjWaitGroup).Wg
-		wg.Done()
-		return value.NewNull()
-	})
-
-	vm.DefineNative("wg_wait", func(args []value.Value) value.Value {
-		if len(args) != 1 {
-			return value.NewNull()
-		}
-		if args[0].Type != value.VAL_WAITGROUP {
-			return value.NewNull()
-		}
-		wg := args[0].Obj.(*value.ObjWaitGroup).Wg
-		wg.Wait()
-		return value.NewNull()
-	})
-
-	vm.DefineNative("to_str", func(args []value.Value) value.Value {
-		if len(args) != 1 {
-			// Should return error or empty?
-			return value.NewString("")
-		}
-		if args[0].Type == value.VAL_BYTES {
-			return value.NewString(args[0].Obj.(string))
-		}
-		return value.NewString(args[0].String())
-	})
-	vm.DefineNative("to_int", func(args []value.Value) value.Value {
-		if len(args) != 1 {
-			return value.NewInt(0)
-		}
-		v := args[0]
-		if v.Type == value.VAL_INT {
-			return value.NewInt(v.AsInt)
-		}
-		if v.Type == value.VAL_FLOAT {
-			return value.NewInt(int64(v.AsFloat))
-		}
-		if v.Type == value.VAL_OBJ {
-			if s, ok := v.Obj.(string); ok {
-				if i, err := strconv.ParseInt(s, 10, 64); err == nil {
-					return value.NewInt(i)
-				}
-				if f, err := strconv.ParseFloat(s, 64); err == nil {
-					return value.NewInt(int64(f))
-				}
-			}
-		}
-		return value.NewInt(0)
-	})
-	vm.DefineNative("to_float", func(args []value.Value) value.Value {
-		if len(args) != 1 {
-			return value.NewFloat(0.0)
-		}
-		v := args[0]
-		if v.Type == value.VAL_FLOAT {
-			return value.NewFloat(v.AsFloat)
-		}
-		if v.Type == value.VAL_INT {
-			return value.NewFloat(float64(v.AsInt))
-		}
-		if v.Type == value.VAL_OBJ {
-			if s, ok := v.Obj.(string); ok {
-				if f, err := strconv.ParseFloat(s, 64); err == nil {
-					return value.NewFloat(f)
-				}
-			}
-		}
-		return value.NewFloat(0.0)
-	})
-	vm.DefineNative("time_now_ms", func(args []value.Value) value.Value {
-		return value.NewInt(time.Now().UnixMilli())
-	})
-	vm.DefineNative("time_now", func(args []value.Value) value.Value {
-		return value.NewInt(time.Now().Unix())
-	})
-
-	vm.DefineNative("time_sleep", func(args []value.Value) value.Value {
-		if len(args) != 1 {
-			return value.NewNull()
-		}
-		ms := args[0].AsInt
-		time.Sleep(time.Duration(ms) * time.Millisecond)
-		return value.NewNull()
-	})
-	vm.DefineNative("time_now_datetime", func(args []value.Value) value.Value {
-		// args[0] is DateTime struct def
-		if len(args) < 1 {
-			return value.NewNull()
-		}
-		structDef, ok := args[0].Obj.(*value.ObjStruct)
-		if !ok {
-			return value.NewNull()
-		}
-
-		t := time.Now()
-		inst := value.NewInstance(structDef).Obj.(*value.ObjInstance)
-		inst.Fields["year"] = value.NewInt(int64(t.Year()))
-		inst.Fields["month"] = value.NewInt(int64(t.Month()))
-		inst.Fields["day"] = value.NewInt(int64(t.Day()))
-		inst.Fields["hour"] = value.NewInt(int64(t.Hour()))
-		inst.Fields["minute"] = value.NewInt(int64(t.Minute()))
-		inst.Fields["second"] = value.NewInt(int64(t.Second()))
-		inst.Fields["weekday"] = value.NewInt(int64(t.Weekday()))
-		inst.Fields["yearday"] = value.NewInt(int64(t.YearDay()))
-		inst.Fields["timestamp"] = value.NewInt(t.Unix())
-
-		return value.Value{Type: value.VAL_OBJ, Obj: inst}
-	})
-	vm.DefineNative("time_format", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewString("")
-		}
-		inst, ok := args[0].Obj.(*value.ObjInstance)
-		if !ok {
-			return value.NewString("")
-		}
-
-		// Reconstruct time.Time from fields
-		// Minimal fields: year, month, day, hour, minute, second
-		y := int(inst.Fields["year"].AsInt)
-		m := time.Month(inst.Fields["month"].AsInt)
-		d := int(inst.Fields["day"].AsInt)
-		h := int(inst.Fields["hour"].AsInt)
-		min := int(inst.Fields["minute"].AsInt)
-		s := int(inst.Fields["second"].AsInt)
-
-		t := time.Date(y, m, d, h, min, s, 0, time.Local)
-		return value.NewString(t.Format("2006-01-02 15:04:05"))
-	})
-	vm.DefineNative("time_format_date", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewString("")
-		}
-		inst, ok := args[0].Obj.(*value.ObjInstance)
-		if !ok {
-			return value.NewString("")
-		}
-		y := int(inst.Fields["year"].AsInt)
-		m := time.Month(inst.Fields["month"].AsInt)
-		d := int(inst.Fields["day"].AsInt)
-		t := time.Date(y, m, d, 0, 0, 0, 0, time.Local)
-		return value.NewString(t.Format("2006-01-02"))
-	})
-	vm.DefineNative("time_format_time", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewString("")
-		}
-		inst, ok := args[0].Obj.(*value.ObjInstance)
-		if !ok {
-			return value.NewString("")
-		}
-		h := int(inst.Fields["hour"].AsInt)
-		min := int(inst.Fields["minute"].AsInt)
-		s := int(inst.Fields["second"].AsInt)
-		t := time.Date(0, 1, 1, h, min, s, 0, time.Local)
-		return value.NewString(t.Format("15:04:05"))
-	})
-	vm.DefineNative("time_make_datetime", func(args []value.Value) value.Value {
-		// args: structDef, y, m, d, h, min, s
-		if len(args) < 7 {
-			return value.NewNull()
-		}
-		structDef, ok := args[0].Obj.(*value.ObjStruct)
-		if !ok {
-			return value.NewNull()
-		}
-
-		y := int(args[1].AsInt)
-		m := time.Month(args[2].AsInt)
-		d := int(args[3].AsInt)
-		h := int(args[4].AsInt)
-		min := int(args[5].AsInt)
-		s := int(args[6].AsInt)
-
-		t := time.Date(y, m, d, h, min, s, 0, time.Local)
-
-		inst := value.NewInstance(structDef).Obj.(*value.ObjInstance)
-		inst.Fields["year"] = value.NewInt(int64(t.Year()))
-		inst.Fields["month"] = value.NewInt(int64(t.Month()))
-		inst.Fields["day"] = value.NewInt(int64(t.Day()))
-		inst.Fields["hour"] = value.NewInt(int64(t.Hour()))
-		inst.Fields["minute"] = value.NewInt(int64(t.Minute()))
-		inst.Fields["second"] = value.NewInt(int64(t.Second()))
-		inst.Fields["weekday"] = value.NewInt(int64(t.Weekday()))
-		inst.Fields["yearday"] = value.NewInt(int64(t.YearDay()))
-		inst.Fields["timestamp"] = value.NewInt(t.Unix())
-
-		return value.Value{Type: value.VAL_OBJ, Obj: inst}
-	})
-	vm.DefineNative("time_to_timestamp", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewInt(0)
-		}
-		inst, ok := args[0].Obj.(*value.ObjInstance)
-		if !ok {
-			return value.NewInt(0)
-		}
-
-		val, ok := inst.Fields["timestamp"]
-		if ok {
-			return val
-		}
-		return value.NewInt(0)
-	})
-	vm.DefineNative("time_from_timestamp", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewNull()
-		}
-		ts := args[0].AsInt
-		structDef, ok := args[1].Obj.(*value.ObjStruct)
-		if !ok {
-			return value.NewNull()
-		}
-
-		t := time.Unix(ts, 0)
-		inst := value.NewInstance(structDef).Obj.(*value.ObjInstance)
-		inst.Fields["year"] = value.NewInt(int64(t.Year()))
-		inst.Fields["month"] = value.NewInt(int64(t.Month()))
-		inst.Fields["day"] = value.NewInt(int64(t.Day()))
-		inst.Fields["hour"] = value.NewInt(int64(t.Hour()))
-		inst.Fields["minute"] = value.NewInt(int64(t.Minute()))
-		inst.Fields["second"] = value.NewInt(int64(t.Second()))
-		inst.Fields["weekday"] = value.NewInt(int64(t.Weekday()))
-		inst.Fields["yearday"] = value.NewInt(int64(t.YearDay()))
-		inst.Fields["timestamp"] = value.NewInt(t.Unix())
-
-		return value.Value{Type: value.VAL_OBJ, Obj: inst}
-	})
-	vm.DefineNative("time_diff", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewInt(0)
-		}
-		ts1 := args[0].AsInt
-		ts2 := args[1].AsInt
-		return value.NewInt(ts1 - ts2)
-	})
-	vm.DefineNative("time_add_days", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewInt(0)
-		}
-		ts := args[0].AsInt
-		days := args[1].AsInt
-		return value.NewInt(ts + (days * 86400))
-	})
-	vm.DefineNative("time_before", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewBool(false)
-		}
-		return value.NewBool(args[0].AsInt < args[1].AsInt)
-	})
-	vm.DefineNative("time_after", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewBool(false)
-		}
-		return value.NewBool(args[0].AsInt > args[1].AsInt)
-	})
-	vm.DefineNative("time_is_leap_year", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewBool(false)
-		}
-		year := args[0].AsInt
-		return value.NewBool(year%4 == 0 && (year%100 != 0 || year%400 == 0))
-	})
-	vm.DefineNative("time_days_in_month", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewInt(0)
-		}
-		year := int(args[0].AsInt)
-		month := time.Month(args[1].AsInt)
-		// Trick: go to next month day 0
-		t := time.Date(year, month+1, 0, 0, 0, 0, 0, time.UTC)
-		return value.NewInt(int64(t.Day()))
-	})
-	vm.DefineNative("time_weekday_name", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewString("")
-		}
-		wd := time.Weekday(args[0].AsInt)
-
-		names := []string{
-			"Domingo", "Segunda-feira", "Terça-feira", "Quarta-feira",
-			"Quinta-feira", "Sexta-feira", "Sábado",
-		}
-		if int(wd) >= 0 && int(wd) < len(names) {
-			return value.NewString(names[wd])
-		}
-		return value.NewString(wd.String())
-	})
-	vm.DefineNative("time_month_name", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewString("")
-		}
-		m := time.Month(args[0].AsInt)
-		names := map[time.Month]string{
-			time.January: "Janeiro", time.February: "Fevereiro", time.March: "Março",
-			time.April: "Abril", time.May: "Maio", time.June: "Junho",
-			time.July: "Julho", time.August: "Agosto", time.September: "Setembro",
-			time.October: "Outubro", time.November: "Novembro", time.December: "Dezembro",
-		}
-		if name, ok := names[m]; ok {
-			return value.NewString(name)
-		}
-		return value.NewString(m.String())
-	})
-	vm.DefineNative("io_open", func(args []value.Value) value.Value {
-		// args: path, mode, FileStructDef
-		if len(args) < 3 {
-			return value.NewNull()
-		}
-		path := args[0].String()
-		mode := args[1].String()
-
-		structDef, ok := args[2].Obj.(*value.ObjStruct)
-		if !ok {
-			return value.NewNull()
-		}
-
-		flag := os.O_RDONLY
-		if mode == "w" {
-			flag = os.O_WRONLY | os.O_CREATE | os.O_TRUNC
-		} else if mode == "a" {
-			flag = os.O_WRONLY | os.O_CREATE | os.O_APPEND
-		} else if mode == "rw" || mode == "r+" {
-			flag = os.O_RDWR | os.O_CREATE
-		}
-
-		f, err := os.OpenFile(path, flag, 0644)
-		isOpen := true
-		var fd int64 = 0
-
-		if err != nil {
-			isOpen = false
-		} else {
-			fd = vm.nextFD
-			vm.nextFD++
-			vm.openFiles[fd] = f
-		}
-
-		inst := value.NewInstance(structDef).Obj.(*value.ObjInstance)
-		inst.Fields["fd"] = value.NewInt(fd)
-		inst.Fields["path"] = value.NewString(path)
-		inst.Fields["mode"] = value.NewString(mode)
-		inst.Fields["open"] = value.NewBool(isOpen)
-
-		return value.Value{Type: value.VAL_OBJ, Obj: inst}
-	})
-	vm.DefineNative("io_close", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewNull()
-		}
-		inst, ok := args[0].Obj.(*value.ObjInstance)
-		if !ok {
-			return value.NewNull()
-		}
-
-		fd := inst.Fields["fd"].AsInt
-		if f, exists := vm.openFiles[fd]; exists {
-			f.Close()
-			delete(vm.openFiles, fd)
-			inst.Fields["open"] = value.NewBool(false)
-		}
-		return value.NewNull()
-	})
-	vm.DefineNative("io_write", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewNull()
-		}
-		inst, ok := args[0].Obj.(*value.ObjInstance)
-		if !ok {
-			return value.NewNull()
-		}
-
-		fd := inst.Fields["fd"].AsInt
-		if f, exists := vm.openFiles[fd]; exists {
-			if args[1].Type == value.VAL_BYTES {
-				// Bytes are stored as string in Obj, but treat as raw bytes
-				data := args[1].Obj.(string)
-				f.Write([]byte(data))
-			} else {
-				content := args[1].String()
-				f.WriteString(content)
-			}
-		}
-		return value.NewNull()
-	})
-	vm.DefineNative("io_read", func(args []value.Value) value.Value {
-		// args: fileInst, IOResultStructDef
-		if len(args) < 2 {
-			return value.NewNull()
-		}
-		inst, ok := args[0].Obj.(*value.ObjInstance)
-		if !ok {
-			return value.NewNull()
-		}
-		resStruct, ok := args[1].Obj.(*value.ObjStruct) // IOResult
-		if !ok {
-			return value.NewNull()
-		}
-
-		fd := inst.Fields["fd"].AsInt
-		var contentStr string
-		var errorStr string
-		var isOk bool = false
-
-		if f, exists := vm.openFiles[fd]; exists {
-			// Read all
-			stat, _ := f.Stat()
-			if stat.Size() > 0 {
-				buf := make([]byte, stat.Size())
-				f.Seek(0, 0)
-				n, err := f.Read(buf)
-				if err == nil || (err != nil && n > 0) { // simple read
-					contentStr = string(buf[:n])
-					isOk = true
-				} else {
-					errorStr = err.Error()
-				}
-			} else {
-				isOk = true // empty file
-			}
-		} else {
-			errorStr = "File not open"
-		}
-
-		resInst := value.NewInstance(resStruct).Obj.(*value.ObjInstance)
-		resInst.Fields["ok"] = value.NewBool(isOk)
-		resInst.Fields["data"] = value.NewString(contentStr)
-		resInst.Fields["error"] = value.NewString(errorStr)
-		return value.Value{Type: value.VAL_OBJ, Obj: resInst}
-	})
-
-	vm.DefineNative("io_read_bytes", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewNull()
-		}
-		inst, ok := args[0].Obj.(*value.ObjInstance)
-		if !ok {
-			return value.NewNull()
-		}
-		resStruct, ok := args[1].Obj.(*value.ObjStruct)
-		if !ok {
-			return value.NewNull()
-		}
-
-		fd := inst.Fields["fd"].AsInt
-		var contentBytes []byte
-		var errorStr string
-		var isOk bool = false
-
-		if f, exists := vm.openFiles[fd]; exists {
-			// Read all
-			stat, _ := f.Stat()
-			if stat.Size() > 0 {
-				buf := make([]byte, stat.Size())
-				f.Seek(0, 0)
-				n, err := f.Read(buf)
-				if err == nil || (err != nil && n > 0) {
-					contentBytes = buf[:n]
-					isOk = true
-				} else {
-					errorStr = err.Error()
-				}
-			} else {
-				contentBytes = []byte{}
-				isOk = true
-			}
-		} else {
-			errorStr = "File not open"
-		}
-
-		resInst := value.NewInstance(resStruct).Obj.(*value.ObjInstance)
-		resInst.Fields["ok"] = value.NewBool(isOk)
-		resInst.Fields["data"] = value.NewBytes(string(contentBytes))
-		resInst.Fields["error"] = value.NewString(errorStr)
-		return value.Value{Type: value.VAL_OBJ, Obj: resInst}
-	})
-	vm.DefineNative("io_exists", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewBool(false)
-		}
-		path := args[0].String()
-		_, err := os.Stat(path)
-		return value.NewBool(err == nil)
-	})
-	vm.DefineNative("io_remove", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewBool(false)
-		}
-		path := args[0].String()
-		err := os.Remove(path)
-		return value.NewBool(err == nil)
-	})
-	vm.DefineNative("io_read_lines", func(args []value.Value) value.Value {
-		// args: fileInst, IOLinesResultStructDef
-		if len(args) < 2 {
-			return value.NewNull()
-		}
-		inst, ok := args[0].Obj.(*value.ObjInstance)
-		if !ok {
-			return value.NewNull()
-		}
-		resStruct, ok := args[1].Obj.(*value.ObjStruct) // IOLinesResult
-		if !ok {
-			return value.NewNull()
-		}
-
-		fd := inst.Fields["fd"].AsInt
-		var lines []string
-		var errorStr string
-		var isOk bool = false
-
-		if f, exists := vm.openFiles[fd]; exists {
-			// Read all
-			stat, _ := f.Stat()
-			var contentStr string
-			if stat.Size() > 0 {
-				f.Seek(0, 0)
-				buf := make([]byte, stat.Size())
-				n, err := f.Read(buf)
-				if err == nil || (err != nil && n > 0) {
-					contentStr = string(buf[:n])
-					isOk = true
-				} else {
-					errorStr = err.Error()
-				}
-			} else {
-				isOk = true
-			}
-
-			if isOk {
-				// Split by newlines, handling \r\n and \n
-				// Naive split
-				contentStr = strings.ReplaceAll(contentStr, "\r\n", "\n")
-				lines = strings.Split(contentStr, "\n")
-				// Retain behavior of strings.Split where trailing newline results in an empty string.
-			}
-		} else {
-			errorStr = "File not open"
-		}
-
-		resInst := value.NewInstance(resStruct).Obj.(*value.ObjInstance)
-		resInst.Fields["ok"] = value.NewBool(isOk)
-
-		linesVal := make([]value.Value, len(lines))
-		for i, line := range lines {
-			linesVal[i] = value.NewString(line)
-		}
-		resInst.Fields["data"] = value.NewArray(linesVal)
-
-		resInst.Fields["error"] = value.NewString(errorStr)
-		return value.Value{Type: value.VAL_OBJ, Obj: resInst}
-	})
-	vm.DefineNative("io_stat", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewNull()
-		}
-		path := args[0].String()
-		structDef, ok := args[1].Obj.(*value.ObjStruct)
-		if !ok {
-			return value.NewNull()
-		}
-
-		info, err := os.Stat(path)
-		exists := (err == nil)
-		size := int64(0)
-		isDir := false
-		if exists {
-			size = info.Size()
-			isDir = info.IsDir()
-		}
-
-		inst := value.NewInstance(structDef).Obj.(*value.ObjInstance)
-		inst.Fields["exists"] = value.NewBool(exists)
-		inst.Fields["size"] = value.NewInt(size)
-		inst.Fields["is_dir"] = value.NewBool(isDir)
-
-		return value.Value{Type: value.VAL_OBJ, Obj: inst}
-	})
-	vm.DefineNative("io_mkdir", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewBool(false)
-		}
-		path := args[0].String()
-		err := os.MkdirAll(path, 0755)
-		return value.NewBool(err == nil)
-	})
-
-	vm.DefineNative("time_format_custom", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewString("")
-		}
-		inst, ok := args[0].Obj.(*value.ObjInstance)
-		if !ok {
-			return value.NewString("")
-		}
-		fmtStr := args[1].Obj.(string)
-
-		y := int(inst.Fields["year"].AsInt)
-		m := time.Month(inst.Fields["month"].AsInt)
-		d := int(inst.Fields["day"].AsInt)
-		h := int(inst.Fields["hour"].AsInt)
-		min := int(inst.Fields["minute"].AsInt)
-		s := int(inst.Fields["second"].AsInt)
-		// t := time.Date(y, m, d, h, min, s, 0, time.Local) // Unused in this simple implementation
-
-		// Simplified replacement for strftime
-		// Noxy: %Y=ano, %m=mês, %d=dia, %H=hora, %M=min, %S=seg
-		res := fmtStr
-		res = strings.ReplaceAll(res, "%Y", fmt.Sprintf("%04d", y))
-		res = strings.ReplaceAll(res, "%m", fmt.Sprintf("%02d", m))
-		res = strings.ReplaceAll(res, "%d", fmt.Sprintf("%02d", d))
-		res = strings.ReplaceAll(res, "%H", fmt.Sprintf("%02d", h))
-		res = strings.ReplaceAll(res, "%M", fmt.Sprintf("%02d", min))
-		res = strings.ReplaceAll(res, "%S", fmt.Sprintf("%02d", s))
-
-		return value.NewString(res)
-	})
-	vm.DefineNative("time_parse", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewNull()
-		}
-		str := args[0].Obj.(string)
-		structDef, ok := args[1].Obj.(*value.ObjStruct)
-		if !ok {
-			return value.NewNull()
-		}
-
-		t, err := time.ParseInLocation("2006-01-02 15:04:05", str, time.Local)
-		if err != nil {
-			return value.NewNull()
-		}
-
-		inst := value.NewInstance(structDef).Obj.(*value.ObjInstance)
-		inst.Fields["year"] = value.NewInt(int64(t.Year()))
-		inst.Fields["month"] = value.NewInt(int64(t.Month()))
-		inst.Fields["day"] = value.NewInt(int64(t.Day()))
-		inst.Fields["hour"] = value.NewInt(int64(t.Hour()))
-		inst.Fields["minute"] = value.NewInt(int64(t.Minute()))
-		inst.Fields["second"] = value.NewInt(int64(t.Second()))
-		inst.Fields["weekday"] = value.NewInt(int64(t.Weekday()))
-		inst.Fields["yearday"] = value.NewInt(int64(t.YearDay()))
-		inst.Fields["timestamp"] = value.NewInt(t.Unix())
-
-		return value.Value{Type: value.VAL_OBJ, Obj: inst}
-	})
-	vm.DefineNative("time_parse_date", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewNull()
-		}
-		str := args[0].Obj.(string)
-		structDef, ok := args[1].Obj.(*value.ObjStruct)
-		if !ok {
-			return value.NewNull()
-		}
-
-		t, err := time.ParseInLocation("2006-01-02", str, time.Local)
-		if err != nil {
-			return value.NewNull()
-		}
-
-		inst := value.NewInstance(structDef).Obj.(*value.ObjInstance)
-		inst.Fields["year"] = value.NewInt(int64(t.Year()))
-		inst.Fields["month"] = value.NewInt(int64(t.Month()))
-		inst.Fields["day"] = value.NewInt(int64(t.Day()))
-		inst.Fields["hour"] = value.NewInt(int64(t.Hour()))
-		inst.Fields["minute"] = value.NewInt(int64(t.Minute()))
-		inst.Fields["second"] = value.NewInt(int64(t.Second()))
-		inst.Fields["weekday"] = value.NewInt(int64(t.Weekday()))
-		inst.Fields["yearday"] = value.NewInt(int64(t.YearDay()))
-		inst.Fields["timestamp"] = value.NewInt(t.Unix())
-
-		return value.Value{Type: value.VAL_OBJ, Obj: inst}
-	})
-	vm.DefineNative("time_add_seconds", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewInt(0)
-		}
-		ts := args[0].AsInt
-		secs := args[1].AsInt
-		return value.NewInt(ts + secs)
-	})
-	vm.DefineNative("time_diff_duration", func(args []value.Value) value.Value {
-		if len(args) < 3 {
-			return value.NewNull()
-		}
-		ts1 := args[0].AsInt
-		ts2 := args[1].AsInt
-		structDef, ok := args[2].Obj.(*value.ObjStruct)
-		if !ok {
-			return value.NewNull()
-		}
-
-		diff := ts1 - ts2
-		if diff < 0 {
-			diff = -diff
-		}
-
-		totalSecs := ts1 - ts2
-		absSecs := totalSecs
-		if absSecs < 0 {
-			absSecs = -absSecs
-		}
-
-		days := absSecs / 86400
-		rem := absSecs % 86400
-		hours := rem / 3600
-		rem = rem % 3600
-		mins := rem / 60
-		secs := rem % 60
-
-		inst := value.NewInstance(structDef).Obj.(*value.ObjInstance)
-		inst.Fields["days"] = value.NewInt(days)
-		inst.Fields["hours"] = value.NewInt(hours)
-		inst.Fields["minutes"] = value.NewInt(mins)
-		inst.Fields["seconds"] = value.NewInt(secs)
-		inst.Fields["total_seconds"] = value.NewInt(totalSecs)
-
-		return value.Value{Type: value.VAL_OBJ, Obj: inst}
-	})
-
-	// Strings Module
-	vm.DefineNative("strings_contains", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewBool(false)
-		}
-		return value.NewBool(strings.Contains(args[0].String(), args[1].String()))
-	})
-	vm.DefineNative("strings_starts_with", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewBool(false)
-		}
-		return value.NewBool(strings.HasPrefix(args[0].String(), args[1].String()))
-	})
-	vm.DefineNative("strings_ends_with", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewBool(false)
-		}
-		return value.NewBool(strings.HasSuffix(args[0].String(), args[1].String()))
-	})
-	vm.DefineNative("strings_index_of", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewInt(-1)
-		}
-		return value.NewInt(int64(strings.Index(args[0].String(), args[1].String())))
-	})
-	vm.DefineNative("strings_count", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewInt(0)
-		}
-		return value.NewInt(int64(strings.Count(args[0].String(), args[1].String())))
-	})
-	vm.DefineNative("strings_to_upper", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewString("")
-		}
-		return value.NewString(strings.ToUpper(args[0].String()))
-	})
-	vm.DefineNative("strings_to_lower", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewString("")
-		}
-		return value.NewString(strings.ToLower(args[0].String()))
-	})
-	vm.DefineNative("strings_trim", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewString("")
-		}
-		return value.NewString(strings.TrimSpace(args[0].String()))
-	})
-
-	// Input
-	vm.DefineNative("input", func(args []value.Value) value.Value {
-		// args[0]: prompt (optional)
-		if len(args) > 0 {
-			fmt.Print(args[0].String())
-		}
-		reader := bufio.NewReader(os.Stdin)
-		text, _ := reader.ReadString('\n')
-		// Trim newline (windows \r\n and unix \n)
-		text = strings.TrimRight(text, "\r\n")
-		return value.NewString(text)
-	})
-	vm.DefineNative("strings_reverse", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewString("")
-		}
-		s := args[0].String()
-		runes := []rune(s)
-		for i, j := 0, len(runes)-1; i < j; i, j = i+1, j-1 {
-			runes[i], runes[j] = runes[j], runes[i]
-		}
-		return value.NewString(string(runes))
-	})
-	vm.DefineNative("strings_repeat", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewString("")
-		}
-		return value.NewString(strings.Repeat(args[0].String(), int(args[1].AsInt)))
-	})
-
-	vm.DefineNative("strings_replace", func(args []value.Value) value.Value {
-		if len(args) < 3 {
-			return value.NewString("")
-		}
-		return value.NewString(strings.ReplaceAll(args[0].String(), args[1].String(), args[2].String()))
-	})
-	vm.DefineNative("strings_replace_first", func(args []value.Value) value.Value {
-		if len(args) < 3 {
-			return value.NewString("")
-		}
-		return value.NewString(strings.Replace(args[0].String(), args[1].String(), args[2].String(), 1))
-	})
-	vm.DefineNative("strings_pad_left", func(args []value.Value) value.Value {
-		if len(args) < 3 {
-			return value.NewString("")
-		}
-		s := args[0].String()
-		totalLen := int(args[1].AsInt)
-		padChar := args[2].String()
-		if len(s) >= totalLen {
-			return value.NewString(s)
-		}
-		padding := totalLen - len(s)
-		return value.NewString(strings.Repeat(padChar, padding) + s)
-	})
-	vm.DefineNative("strings_split", func(args []value.Value) value.Value {
-		if len(args) < 3 {
-			return value.NewNull()
-		}
-		s := args[0].String()
-		sep := args[1].String()
-		structDef, ok := args[2].Obj.(*value.ObjStruct)
-		if !ok {
-			return value.NewNull()
-		}
-
-		parts := strings.Split(s, sep)
-
-		inst := value.NewInstance(structDef).Obj.(*value.ObjInstance)
-		inst.Fields["count"] = value.NewInt(int64(len(parts)))
-
-		partValues := make([]value.Value, len(parts))
-		for i, p := range parts {
-			partValues[i] = value.NewString(p)
-		}
-		inst.Fields["parts"] = value.NewArray(partValues)
-
-		return value.Value{Type: value.VAL_OBJ, Obj: inst}
-	})
-	vm.DefineNative("strings_join_count", func(args []value.Value) value.Value {
-		if len(args) < 3 {
-			return value.NewString("")
-		}
-		arrVal := args[0]
-		sep := args[1].String()
-		count := int(args[2].AsInt)
-
-		if arrVal.Type == value.VAL_OBJ {
-			if arr, ok := arrVal.Obj.(*value.ObjArray); ok {
-				var parts []string
-				max := len(arr.Elements)
-				if count < max {
-					max = count
-				}
-				for i := 0; i < max; i++ {
-					parts = append(parts, arr.Elements[i].String())
-				}
-				return value.NewString(strings.Join(parts, sep))
-			}
-		}
-		return value.NewString("")
-	})
-	vm.DefineNative("ord", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewInt(0)
-		}
-		s := args[0].String()
-		if len(s) == 0 {
-			return value.NewInt(0)
-		}
-		return value.NewInt(int64(s[0]))
-	})
-	vm.DefineNative("strings_contains", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewBool(false)
-		}
-		s := args[0].String()
-		substr := args[1].String()
-		return value.NewBool(strings.Contains(s, substr))
-	})
-	vm.DefineNative("strings_replace", func(args []value.Value) value.Value {
-		if len(args) < 3 {
-			return value.NewString("")
-		}
-		s := args[0].String()
-		old := args[1].String()
-		new := args[2].String()
-		return value.NewString(strings.ReplaceAll(s, old, new))
-	})
-	vm.DefineNative("strings_substring", func(args []value.Value) value.Value {
-		// args: string, start, length
-		if len(args) < 3 {
-			return value.NewString("")
-		}
-		s := args[0].String()
-		runes := []rune(s)
-		start := int(args[1].AsInt)
-		length := int(args[2].AsInt)
-
-		if start < 0 {
-			start = 0
-		}
-		if start >= len(runes) {
-			return value.NewString("")
-		}
-
-		end := start + length
-		if end > len(runes) {
-			end = len(runes)
-		}
-
-		return value.NewString(string(runes[start:end]))
-	})
-	vm.DefineNative("strings_is_empty", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewBool(true)
-		}
-		return value.NewBool(len(args[0].String()) == 0)
-	})
-	vm.DefineNative("strings_is_digit", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewBool(false)
-		}
-		s := args[0].String()
-		if len(s) == 0 {
-			return value.NewBool(false)
-		}
-		for _, r := range s {
-			if r < '0' || r > '9' {
-				return value.NewBool(false)
-			}
-		}
-		return value.NewBool(true)
-	})
-	vm.DefineNative("strings_is_alpha", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewBool(false)
-		}
-		s := args[0].String()
-		if len(s) == 0 {
-			return value.NewBool(false)
-		}
-		for _, r := range s {
-			if (r < 'a' || r > 'z') && (r < 'A' || r > 'Z') {
-				return value.NewBool(false)
-			}
-		}
-		return value.NewBool(true)
-	})
-	vm.DefineNative("strings_is_alnum", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewBool(false)
-		}
-		s := args[0].String()
-		if len(s) == 0 {
-			return value.NewBool(false)
-		}
-		for _, r := range s {
-			isDigit := r >= '0' && r <= '9'
-			isAlpha := (r >= 'a' && r <= 'z') || (r >= 'A' && r <= 'Z')
-			if !isDigit && !isAlpha {
-				return value.NewBool(false)
-			}
-		}
-		return value.NewBool(true)
-	})
-	vm.DefineNative("strings_is_space", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewBool(false)
-		}
-		s := args[0].String()
-		if len(s) == 0 {
-			return value.NewBool(false)
-		}
-		for _, r := range s {
-			if r != ' ' && r != '\t' && r != '\n' && r != '\r' {
-				return value.NewBool(false)
-			}
-		}
-		return value.NewBool(true)
-	})
-	vm.DefineNative("strings_char_at", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewString("")
-		}
-		s := args[0].String()
-		runes := []rune(s)
-		idx := int(args[1].AsInt)
-		if idx < 0 || idx >= len(runes) {
-			return value.NewString("")
-		}
-		return value.NewString(string(runes[idx]))
-	})
-	vm.DefineNative("strings_from_char_code", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewString("")
-		}
-		return value.NewString(string(rune(args[0].AsInt)))
-	})
-
-	// Crypto Module - Native implementations for cryptographic operations
-	vm.DefineNative("crypto_random_bytes", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewNull()
-		}
-		n := int(args[0].AsInt)
-		if n <= 0 {
-			return value.NewNull()
-		}
-
-		bytes := make([]byte, n)
-		if _, err := rand.Read(bytes); err != nil {
-			return value.NewNull()
-		}
-
-		return value.NewBytes(string(bytes))
-	})
-
-	vm.DefineNative("crypto_pbkdf2_sha256", func(args []value.Value) value.Value {
-		// args: (senha: string, salt: bytes, iteracoes: int, tamanho: int)
-		if len(args) < 4 {
-			return value.NewNull()
-		}
-
-		senha := args[0].String()
-		var salt []byte
-		if args[1].Type == value.VAL_BYTES {
-			salt = []byte(args[1].Obj.(string))
-		} else {
-			salt = []byte(args[1].String())
-		}
-		iteracoes := int(args[2].AsInt)
-		tamanho := int(args[3].AsInt)
-
-		if iteracoes <= 0 || tamanho <= 0 {
-			return value.NewNull()
-		}
-
-		chave, err := pbkdf2.Key(sha256.New, senha, salt, iteracoes, tamanho)
-		if err != nil {
-			return value.NewNull()
-		}
-		return value.NewBytes(string(chave))
-	})
-
-	vm.DefineNative("crypto_aes256_gcm_encrypt", func(args []value.Value) value.Value {
-		// args: (chave: bytes, texto: bytes) -> bytes (nonce + ciphertext + tag)
-		if len(args) < 2 {
-			return value.NewNull()
-		}
-
-		var chave []byte
-		if args[0].Type == value.VAL_BYTES {
-			chave = []byte(args[0].Obj.(string))
-		} else {
-			chave = []byte(args[0].String())
-		}
-
-		var texto []byte
-		if args[1].Type == value.VAL_BYTES {
-			texto = []byte(args[1].Obj.(string))
-		} else {
-			texto = []byte(args[1].String())
-		}
-
-		// Validate key size (32 bytes for AES-256)
-		if len(chave) != 32 {
-			return value.NewNull()
-		}
-
-		block, err := aes.NewCipher(chave)
-		if err != nil {
-			return value.NewNull()
-		}
-
-		gcm, err := cipher.NewGCM(block)
-		if err != nil {
-			return value.NewNull()
-		}
-
-		// Generate random nonce (12 bytes for GCM)
-		nonce := make([]byte, gcm.NonceSize())
-		if _, err := rand.Read(nonce); err != nil {
-			return value.NewNull()
-		}
-
-		// Seal: result = nonce + ciphertext + tag
-		resultado := gcm.Seal(nonce, nonce, texto, nil)
-		return value.NewBytes(string(resultado))
-	})
-
-	vm.DefineNative("crypto_aes256_gcm_decrypt", func(args []value.Value) value.Value {
-		// args: (chave: bytes, dados: bytes) -> bytes (plaintext)
-		if len(args) < 2 {
-			return value.NewNull()
-		}
-
-		var chave []byte
-		if args[0].Type == value.VAL_BYTES {
-			chave = []byte(args[0].Obj.(string))
-		} else {
-			chave = []byte(args[0].String())
-		}
-
-		var dados []byte
-		if args[1].Type == value.VAL_BYTES {
-			dados = []byte(args[1].Obj.(string))
-		} else {
-			dados = []byte(args[1].String())
-		}
-
-		// Validate key size (32 bytes for AES-256)
-		if len(chave) != 32 {
-			return value.NewNull()
-		}
-
-		block, err := aes.NewCipher(chave)
-		if err != nil {
-			return value.NewNull()
-		}
-
-		gcm, err := cipher.NewGCM(block)
-		if err != nil {
-			return value.NewNull()
-		}
-
-		nonceSize := gcm.NonceSize()
-		if len(dados) < nonceSize {
-			return value.NewNull()
-		}
-
-		// Separate nonce from ciphertext
-		nonce := dados[:nonceSize]
-		ciphertext := dados[nonceSize:]
-
-		// Open: decrypt and validate authentication
-		texto, err := gcm.Open(nil, nonce, ciphertext, nil)
-		if err != nil {
-			return value.NewNull()
-		}
-
-		return value.NewBytes(string(texto))
-	})
-
-	// Sys Module
-	vm.DefineNative("sys_os", func(args []value.Value) value.Value {
-		return value.NewString(runtime.GOOS)
-	})
-
-	vm.DefineNative("sys_exec", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewNull()
-		}
-		cmdStr := args[0].String()
-		structDef, ok := args[1].Obj.(*value.ObjStruct)
-		if !ok {
-			return value.NewNull()
-		}
-
-		var cmd *exec.Cmd
-		if os.PathSeparator == '\\' {
-			cmd = exec.Command("cmd", "/C", cmdStr)
-		} else {
-			cmd = exec.Command("sh", "-c", cmdStr)
-		}
-
-		cmd.Stdout = os.Stdout
-		cmd.Stderr = os.Stderr
-
-		err := cmd.Run()
-		exitCode := 0
-		okVal := true
-
-		var outputStr string = "" // No captured output for sys_exec
-
-		if err != nil {
-			okVal = false
-			if exitErr, ok := err.(*exec.ExitError); ok {
-				exitCode = exitErr.ExitCode()
-			} else {
-				exitCode = 1
-			}
-		}
-
-		inst := value.NewInstance(structDef).Obj.(*value.ObjInstance)
-		inst.Fields["exit_code"] = value.NewInt(int64(exitCode))
-		inst.Fields["output"] = value.NewString(outputStr)
-		inst.Fields["ok"] = value.NewBool(okVal)
-
-		return value.Value{Type: value.VAL_OBJ, Obj: inst}
-	})
-
-	vm.DefineNative("sys_exec_output", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewNull()
-		}
-		cmdStr := args[0].String()
-		structDef, ok := args[1].Obj.(*value.ObjStruct)
-		if !ok {
-			return value.NewNull()
-		}
-
-		var cmd *exec.Cmd
-		if os.PathSeparator == '\\' {
-			cmd = exec.Command("cmd", "/C", cmdStr)
-		} else {
-			cmd = exec.Command("sh", "-c", cmdStr)
-		}
-
-		outBytes, err := cmd.CombinedOutput()
-		outputStr := string(outBytes)
-
-		// OK indicates execution completion, regardless of exit code.
-		okVal := true
-		exitCode := 0
-
-		if err != nil {
-			if exitErr, ok := err.(*exec.ExitError); ok {
-				exitCode = exitErr.ExitCode()
-			} else {
-				exitCode = 1
-			}
-			okVal = false
-		} else {
-			okVal = true
-		}
-
-		inst := value.NewInstance(structDef).Obj.(*value.ObjInstance)
-		inst.Fields["exit_code"] = value.NewInt(int64(exitCode))
-		inst.Fields["output"] = value.NewString(strings.TrimSpace(outputStr))
-		inst.Fields["ok"] = value.NewBool(okVal)
-
-		return value.Value{Type: value.VAL_OBJ, Obj: inst}
-	})
-
-	vm.DefineNative("sys_load_plugin", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewBool(false)
-		}
-		name := args[0].String()
-		cmdName := args[1].String()
-
-		// Intelligent Path Search
-		var cmdPath string
-		found := false
-
-		// 1. Check absolute path or PATH override
-		if filepath.IsAbs(cmdName) {
-			if _, err := os.Stat(cmdName); err == nil {
-				cmdPath = cmdName
-				found = true
-			}
-		} else {
-			// 2. Check path provided directly (PATH lookup)
-			if path, err := exec.LookPath(cmdName); err == nil {
-				cmdPath = path
-				found = true
-			}
-		}
-
-		// 3. Check Current Working Directory (explicitly)
-		if !found {
-			cwd, _ := os.Getwd()
-			localPath := filepath.Join(cwd, cmdName)
-			// Add .exe on Windows if not present
-			if runtime.GOOS == "windows" && !strings.HasSuffix(localPath, ".exe") {
-				localPath += ".exe"
-			}
-			if _, err := os.Stat(localPath); err == nil {
-				cmdPath = localPath
-				found = true
-			}
-		}
-
-		// 4. Check noxy_libs recursively (Depth restricted)
-		if !found {
-			cwd, _ := os.Getwd()
-			libsDir := filepath.Join(cwd, "noxy_libs")
-			filepath.Walk(libsDir, func(path string, info os.FileInfo, err error) error {
-				if found {
-					return filepath.SkipDir // Stop if found
-				}
-				if err != nil {
-					return nil // Ignore errors
-				}
-				if info.IsDir() {
-					if info.Name() == ".git" {
-						return filepath.SkipDir
-					}
-					return nil
-				}
-
-				fname := info.Name()
-				isMatch := fname == cmdName
-				if runtime.GOOS == "windows" {
-					isMatch = fname == cmdName || fname == cmdName+".exe"
-				}
-
-				if isMatch {
-					cmdPath = path
-					found = true
-					return filepath.SkipDir // Abort walk
-				}
-				return nil
-			})
-		}
-
-		if !found {
-			fmt.Printf("Plugin Load Error: command not found: %s\n", cmdName)
-			return value.NewBool(false)
-		}
-
-		client, err := plugin.LoadPlugin(name, cmdPath)
-		if err != nil {
-			fmt.Printf("Plugin Load Error: failed to load plugin: %v\n", err)
-			return value.NewBool(false)
-		}
-
-		// Define Native dynamically
-		nativeName := name + "_request" // e.g. dynamodb_request
-		vm.DefineNative(nativeName, func(args []value.Value) value.Value {
-			if len(args) < 1 {
-				return value.NewNull()
-			}
-			method := args[0].String()
-			params := args[1:]
-			return client.Call(method, params)
-		})
-
-		return value.NewBool(true)
-	})
-
-	vm.DefineNative("sys_getenv", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewNull()
-		}
-		key := args[0].String()
-		structDef, ok := args[1].Obj.(*value.ObjStruct)
-		if !ok {
-			return value.NewNull()
-		}
-
-		val, found := os.LookupEnv(key)
-
-		inst := value.NewInstance(structDef).Obj.(*value.ObjInstance)
-		inst.Fields["value"] = value.NewString(val)
-		inst.Fields["ok"] = value.NewBool(found)
-
-		return value.Value{Type: value.VAL_OBJ, Obj: inst}
-	})
-
-	vm.DefineNative("sys_setenv", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewBool(false)
-		}
-		key := args[0].String()
-		val := args[1].String()
-		err := os.Setenv(key, val)
-		return value.NewBool(err == nil)
-	})
-
-	vm.DefineNative("sys_getcwd", func(args []value.Value) value.Value {
-		dir, err := os.Getwd()
-		if err != nil {
-			return value.NewString("")
-		}
-		return value.NewString(dir)
-	})
-
-	vm.DefineNative("sys_argv", func(args []value.Value) value.Value {
-		// Convert os.Args to string[]
-		vals := make([]value.Value, len(os.Args))
-		for i, a := range os.Args {
-			vals[i] = value.NewString(a)
-		}
-		return value.NewArray(vals)
-	})
-
-	vm.DefineNative("sys_sleep", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewNull()
-		}
-		ms := args[0].AsInt
-		time.Sleep(time.Duration(ms) * time.Millisecond)
-		return value.NewNull()
-	})
-
-	vm.DefineNative("sys_exit", func(args []value.Value) value.Value {
-		code := 0
-		if len(args) > 0 {
-			code = int(args[0].AsInt)
-		}
-		os.Exit(code)
-		return value.NewNull()
-	})
-
-	vm.DefineNative("length", func(args []value.Value) value.Value {
-		if len(args) != 1 {
-			return value.NewInt(0)
-		}
-		arg := args[0]
-		if arg.Type == value.VAL_BYTES {
-			if str, ok := arg.Obj.(string); ok {
-				return value.NewInt(int64(len(str)))
-			}
-		}
-		if arg.Type == value.VAL_OBJ {
-			if str, ok := arg.Obj.(string); ok {
-				return value.NewInt(int64(utf8.RuneCountInString(str)))
-			}
-			if arr, ok := arg.Obj.(*value.ObjArray); ok {
-				return value.NewInt(int64(len(arr.Elements)))
-			}
-			if mp, ok := arg.Obj.(*value.ObjMap); ok {
-				return value.NewInt(int64(len(mp.Data)))
-			}
-		}
-		return value.NewInt(0)
-	})
-
-	vm.DefineNative("keys", func(args []value.Value) value.Value {
-		if len(args) != 1 {
-			return value.NewArray(nil)
-		}
-		mapVal := args[0]
-		if mapVal.Type == value.VAL_OBJ {
-			if m, ok := mapVal.Obj.(*value.ObjMap); ok {
-				keys := make([]value.Value, 0, len(m.Data))
-				for k := range m.Data {
-					if kInt, ok := k.(int64); ok {
-						keys = append(keys, value.NewInt(kInt))
-					} else if kStr, ok := k.(string); ok {
-						keys = append(keys, value.NewString(kStr))
-					}
-				}
-				return value.NewArray(keys)
-			}
-		}
-		return value.NewArray(nil)
-	})
-
-	deleteSignature := value.NativeSignature{
-		Arity: 2,
-		Params: []value.ParamInfo{
-			{IsRef: true, TypeName: "ref map"},
-			{IsRef: false, TypeName: "any"},
-		},
-		ReturnType: "void",
-	}
-	vm.DefineNativeWithSignature("delete", deleteSignature, func(args []value.Value) value.Value {
-		if len(args) != 2 {
-			return value.NewNull()
-		}
-		mapVal, err := vm.resolveReferenceValue(args[0])
-		if err != nil {
-			return value.NewNull()
-		}
-		keyVal := args[1]
-		if mapVal.Type == value.VAL_OBJ {
-			if m, ok := mapVal.Obj.(*value.ObjMap); ok {
-				var key interface{}
-				if keyVal.Type == value.VAL_INT {
-					key = keyVal.AsInt
-				} else if keyVal.Type == value.VAL_OBJ {
-					if str, ok := keyVal.Obj.(string); ok {
-						key = str
-					}
-				}
-				if key != nil {
-					delete(m.Data, key)
-				}
-			}
-		}
-		return value.NewNull()
-	})
-	appendSignature := value.NativeSignature{
-		Arity: 2,
-		Params: []value.ParamInfo{
-			{IsRef: true, TypeName: "ref array"},
-		},
-		ReturnType: "void",
-	}
-	vm.DefineNativeWithSignature("append", appendSignature, func(args []value.Value) value.Value {
-		if len(args) != 2 {
-			return value.NewNull()
-		}
-		targetRef, _ := args[0].Obj.(*value.ObjRef)
-		arrVal, err := vm.resolveReferenceValue(args[0])
-		if err != nil {
-			return value.NewNull()
-		}
-		item := args[1]
-		if !vm.appendItemCompatible(targetRef, item) {
-			return value.NewNull()
-		}
-		if arrVal.Type == value.VAL_OBJ {
-			if arr, ok := arrVal.Obj.(*value.ObjArray); ok {
-				arr.Elements = append(arr.Elements, item)
-			}
-		}
-		return value.NewNull()
-	})
-	popSignature := value.NativeSignature{
-		Arity: 1,
-		Params: []value.ParamInfo{
-			{IsRef: true, TypeName: "ref array"},
-		},
-		ReturnType: "any",
-	}
-	vm.DefineNativeWithSignature("pop", popSignature, func(args []value.Value) value.Value {
-		if len(args) != 1 {
-			return value.NewNull()
-		}
-		arrVal, err := vm.resolveReferenceValue(args[0])
-		if err != nil {
-			return value.NewNull()
-		}
-		if arrVal.Type == value.VAL_OBJ {
-			if arr, ok := arrVal.Obj.(*value.ObjArray); ok {
-				if len(arr.Elements) == 0 {
-					return value.NewNull()
-				}
-				val := arr.Elements[len(arr.Elements)-1]
-				arr.Elements = arr.Elements[:len(arr.Elements)-1]
-				return val
-			}
-		}
-		return value.NewNull()
-	})
-	vm.DefineNative("slice", func(args []value.Value) value.Value {
-		if len(args) < 3 {
-			return value.NewNull()
-		}
-		seq := args[0]
-		start := int(args[1].AsInt)
-		end := int(args[2].AsInt)
-
-		// Clamp logic helper
-		clamp := func(idx, length int) int {
-			if idx < 0 {
-				return 0
-			}
-			if idx > length {
-				return length
-			}
-			return idx
-		}
-
-		switch seq.Type {
-		case value.VAL_OBJ:
-			if str, ok := seq.Obj.(string); ok {
-				runes := []rune(str)
-				start = clamp(start, len(runes))
-				end = clamp(end, len(runes))
-				if start > end {
-					return value.NewString("")
-				}
-				return value.NewString(string(runes[start:end]))
-			}
-			if arr, ok := seq.Obj.(*value.ObjArray); ok {
-				start = clamp(start, len(arr.Elements))
-				end = clamp(end, len(arr.Elements))
-				if start > end {
-					return value.NewArray(nil)
-				}
-
-				newElems := make([]value.Value, end-start)
-				copy(newElems, arr.Elements[start:end])
-				return value.NewArray(newElems)
-			}
-		case value.VAL_BYTES:
-			if str, ok := seq.Obj.(string); ok {
-				// Bytes stored as string
-				start = clamp(start, len(str))
-				end = clamp(end, len(str))
-				if start > end {
-					return value.NewBytes("")
-				}
-				return value.NewBytes(str[start:end])
-			}
-		}
-		return value.NewNull()
-	})
-	vm.DefineNative("contains", func(args []value.Value) value.Value {
-		if len(args) != 2 {
-			return value.NewBool(false)
-		}
-		arrVal := args[0]
-		target := args[1]
-		if arrVal.Type == value.VAL_OBJ {
-			if arr, ok := arrVal.Obj.(*value.ObjArray); ok {
-				for _, el := range arr.Elements {
-					if valuesEqual(el, target) {
-						return value.NewBool(true)
-					}
-				}
-			}
-		}
-		return value.NewBool(false)
-	})
-	vm.DefineNative("has_key", func(args []value.Value) value.Value {
-		if len(args) != 2 {
-			return value.NewBool(false)
-		}
-		mapVal := args[0]
-		keyVal := args[1]
-		if mapVal.Type == value.VAL_OBJ {
-			if mapObj, ok := mapVal.Obj.(*value.ObjMap); ok {
-				var key interface{}
-				if keyVal.Type == value.VAL_INT {
-					key = keyVal.AsInt
-				} else if keyVal.Type == value.VAL_OBJ {
-					if str, ok := keyVal.Obj.(string); ok {
-						key = str
-					} else {
-						return value.NewBool(false)
-					}
-				} else {
-					return value.NewBool(false)
-				}
-				_, ok := mapObj.Data[key]
-				return value.NewBool(ok)
-			}
-		}
-		return value.NewBool(false)
-	})
-	vm.DefineNative("to_bytes", func(args []value.Value) value.Value {
-		if len(args) != 1 {
-			return value.NewBytes("")
-		}
-		arg := args[0]
-		switch arg.Type {
-		case value.VAL_OBJ:
-			if str, ok := arg.Obj.(string); ok {
-				return value.NewBytes(str)
-			}
-			if arr, ok := arg.Obj.(*value.ObjArray); ok {
-				// Array of ints -> bytes
-				bs := make([]byte, len(arr.Elements))
-				for i, el := range arr.Elements {
-					if el.Type == value.VAL_INT {
-						bs[i] = byte(el.AsInt)
-					}
-				}
-				return value.NewBytes(string(bs))
-			}
-		case value.VAL_INT:
-			// Single int to single byte
-			return value.NewBytes(string([]byte{byte(arg.AsInt)}))
-		}
-		return value.NewBytes("")
-	})
-
-	// Net Native Functions
-	vm.DefineNative("net_listen", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewNull()
-		}
-		host := args[0].String()
-		port := int(args[1].AsInt)
-		addr := fmt.Sprintf("%s:%d", host, port)
-
-		listener, err := net.Listen("tcp", addr)
-		if err != nil {
-			// Return Socket with open=false
-			socketFields := map[string]value.Value{
-				"fd":   value.NewInt(-1),
-				"addr": value.NewString(host),
-				"port": value.NewInt(int64(port)),
-				"open": value.NewBool(false),
-			}
-			return value.NewMapWithData(socketFields)
-		}
-
-		vm.shared.NetLock.Lock()
-		id := vm.shared.NextNetID
-		vm.shared.NextNetID++
-		vm.shared.NetListeners[id] = listener
-		vm.shared.NetLock.Unlock()
-
-		socketFields := map[string]value.Value{
-			"fd":   value.NewInt(int64(id)),
-			"addr": value.NewString(host),
-			"port": value.NewInt(int64(port)),
-			"open": value.NewBool(true),
-		}
-		return value.NewMapWithData(socketFields)
-	})
-
-	vm.DefineNative("net_accept", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewNull()
-		}
-		sockMap, ok := args[0].Obj.(*value.ObjMap)
-		if !ok {
-			return value.NewNull()
-		}
-		fdVal, exists := sockMap.Data["fd"]
-		if !exists {
-			return value.NewNull()
-		}
-		fd := int(fdVal.AsInt)
-
-		vm.shared.NetLock.Lock()
-		listener, ok := vm.shared.NetListeners[fd]
-		vm.shared.NetLock.Unlock()
-
-		if !ok {
-			socketFields := map[string]value.Value{
-				"fd":   value.NewInt(-1),
-				"addr": value.NewString(""),
-				"port": value.NewInt(0),
-				"open": value.NewBool(false),
-			}
-			return value.NewMapWithData(socketFields)
-		}
-
-		// Check buffered connection from select
-		var conn net.Conn
-		var err error
-
-		if bufferedConn, ok := vm.netBufferedConns[fd]; ok {
-			conn = bufferedConn
-			delete(vm.netBufferedConns, fd)
-		} else {
-			// Accept blocks. Lock is released above.
-			conn, err = listener.Accept()
-		}
-
-		if err != nil {
-			socketFields := map[string]value.Value{
-				"fd":   value.NewInt(-1),
-				"addr": value.NewString(""),
-				"port": value.NewInt(0),
-				"open": value.NewBool(false),
-			}
-			return value.NewMapWithData(socketFields)
-		}
-
-		vm.shared.NetLock.Lock()
-		id := vm.shared.NextNetID
-		vm.shared.NextNetID++
-		vm.shared.NetConns[id] = conn
-		vm.shared.NetLock.Unlock()
-
-		remoteAddr := conn.RemoteAddr().String()
-		socketFields := map[string]value.Value{
-			"fd":   value.NewInt(int64(id)),
-			"addr": value.NewString(remoteAddr),
-			"port": value.NewInt(0),
-			"open": value.NewBool(true),
-		}
-		return value.NewMapWithData(socketFields)
-	})
-
-	vm.DefineNative("net_connect", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewNull()
-		}
-		host := args[0].String()
-		port := int(args[1].AsInt)
-		addr := fmt.Sprintf("%s:%d", host, port)
-
-		conn, err := net.DialTimeout("tcp", addr, 5*time.Second)
-		if err != nil {
-			socketFields := map[string]value.Value{
-				"fd":   value.NewInt(-1),
-				"addr": value.NewString(host),
-				"port": value.NewInt(int64(port)),
-				"open": value.NewBool(false),
-			}
-			return value.NewMapWithData(socketFields)
-		}
-
-		vm.shared.NetLock.Lock()
-		id := vm.shared.NextNetID
-		vm.shared.NextNetID++
-		vm.shared.NetConns[id] = conn
-		vm.shared.NetLock.Unlock()
-
-		socketFields := map[string]value.Value{
-			"fd":   value.NewInt(int64(id)),
-			"addr": value.NewString(host),
-			"port": value.NewInt(int64(port)),
-			"open": value.NewBool(true),
-		}
-		return value.NewMapWithData(socketFields)
-	})
-
-	vm.DefineNative("net_recv", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewNull()
-		}
-		sockMap, ok := args[0].Obj.(*value.ObjMap)
-		if !ok {
-			return value.NewNull()
-		}
-		fdVal, _ := sockMap.Data["fd"]
-		fd := int(fdVal.AsInt)
-		size := int(args[1].AsInt)
-
-		vm.shared.NetLock.Lock()
-		conn, ok := vm.shared.NetConns[fd]
-		vm.shared.NetLock.Unlock()
-
-		if !ok {
-			resultFields := map[string]value.Value{
-				"ok":    value.NewBool(false),
-				"data":  value.NewBytes(""),
-				"count": value.NewInt(0),
-				"error": value.NewString("invalid socket"),
-			}
-			return value.NewMapWithData(resultFields)
-		}
-
-		var n int
-		buf := make([]byte, size)
-
-		// Check buffered data from select
-		if buffered, ok := vm.netBufferedData[fd]; ok {
-			// Copy buffered data
-			copy(buf, buffered)
-			n = len(buffered)
-			delete(vm.netBufferedData, fd)
-		}
-
-		// Try to read more if space available
-		if n < size {
-			// Blocking read (no deadline)
-			n2, err2 := conn.Read(buf[n:])
-			if n2 > 0 {
-				n += n2
-			}
-
-			// Ignore timeout errors if we have at least some data
-			if err2 != nil {
-				if n == 0 {
-					// Only return error if we really got nothing
-					if err2 != nil && n2 == 0 {
-						if err2 == io.EOF {
-							// Return ok=true, count=0 for EOF
-							resultFields := map[string]value.Value{
-								"ok":    value.NewBool(true),
-								"data":  value.NewBytes(""),
-								"count": value.NewInt(0),
-								"error": value.NewString(""),
-							}
-							return value.NewMapWithData(resultFields)
-						}
-						resultFields := map[string]value.Value{
-							"ok":    value.NewBool(false),
-							"data":  value.NewBytes(""),
-							"count": value.NewInt(0),
-							"error": value.NewString(err2.Error()),
-						}
-						return value.NewMapWithData(resultFields)
-					}
-				}
-			}
-		}
-
-		resultFields := map[string]value.Value{
-			"ok":    value.NewBool(true),
-			"data":  value.NewBytes(string(buf[:n])),
-			"count": value.NewInt(int64(n)),
-			"error": value.NewString(""),
-		}
-		return value.NewMapWithData(resultFields)
-	})
-
-	vm.DefineNative("net_send", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewNull()
-		}
-		sockMap, ok := args[0].Obj.(*value.ObjMap)
-		if !ok {
-			fmt.Printf("DEBUG: net_send args[0] not map: %T %v\n", args[0].Obj, args[0].Obj)
-			return value.NewNull()
-		}
-		fdVal, _ := sockMap.Data["fd"]
-		fd := int(fdVal.AsInt)
-		var data string
-		if args[1].Type == value.VAL_BYTES {
-			data = args[1].Obj.(string)
-		} else {
-			data = args[1].String()
-		}
-
-		vm.shared.NetLock.Lock()
-		conn, ok := vm.shared.NetConns[fd]
-		vm.shared.NetLock.Unlock()
-
-		if !ok {
-			resultFields := map[string]value.Value{
-				"ok":    value.NewBool(false),
-				"data":  value.NewBytes(""),
-				"count": value.NewInt(0),
-				"error": value.NewString("invalid socket"),
-			}
-			return value.NewMapWithData(resultFields)
-		}
-
-		n, err := conn.Write([]byte(data))
-		if err != nil {
-			resultFields := map[string]value.Value{
-				"ok":    value.NewBool(false),
-				"data":  value.NewBytes(""),
-				"count": value.NewInt(0),
-				"error": value.NewString(err.Error()),
-			}
-			return value.NewMapWithData(resultFields)
-		}
-
-		resultFields := map[string]value.Value{
-			"ok":    value.NewBool(true),
-			"data":  value.NewBytes(""),
-			"count": value.NewInt(int64(n)),
-			"error": value.NewString(""),
-		}
-		return value.NewMapWithData(resultFields)
-	})
-
-	vm.DefineNative("net_close", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewNull()
-		}
-
-		var fd int
-		// Check if arg is int (new style) or map (old style compatibility if needed, but we changed net.nx)
-		if args[0].Type == value.VAL_INT {
-			fd = int(args[0].AsInt)
-		} else if args[0].Type == value.VAL_OBJ {
-			// Fallback for old calls? Or just error.
-			if sockMap, ok := args[0].Obj.(*value.ObjMap); ok {
-				if fdVal, found := sockMap.Data["fd"]; found {
-					fd = int(fdVal.AsInt)
-				}
-			}
-		} else {
-			return value.NewNull()
-		}
-
-		vm.shared.NetLock.Lock()
-		defer vm.shared.NetLock.Unlock()
-
-		// Try closing as listener
-		if listener, ok := vm.shared.NetListeners[fd]; ok {
-			listener.Close()
-			delete(vm.shared.NetListeners, fd)
-			return value.NewNull()
-		}
-
-		// Try closing as connection
-		if conn, ok := vm.shared.NetConns[fd]; ok {
-			conn.Close()
-			delete(vm.shared.NetConns, fd)
-		}
-
-		return value.NewNull()
-	})
-
-	vm.DefineNative("net_setblocking", func(args []value.Value) value.Value {
-		// For TCP in Go, blocking is handled at a different level
-		// This is a no-op for now, as Go handles timeouts via SetDeadline
-		return value.NewNull()
-	})
-
-	vm.DefineNative("net_select", func(args []value.Value) value.Value {
-		// args: read, write (ignored), err (ignored), timeout
-		if len(args) < 4 {
-			return value.NewNull() // Or error map
-		}
-
-		timeoutMs := int(args[3].AsInt)
-		// Minimum 1ms to allow polling
-		if timeoutMs < 1 {
-			timeoutMs = 1
-		}
-
-		// Prepare Result Data
-		readyRead := make([]value.Value, 0)
-
-		// Parse Read Set
-		readArrVal := args[0]
-		if readArrVal.Type == value.VAL_OBJ {
-			if arr, ok := readArrVal.Obj.(*value.ObjArray); ok {
-				for _, el := range arr.Elements {
-					if el.Type == value.VAL_OBJ { // Check if socket (Map or Instance)
-						// Extract FD
-						var fd int64 = -1
-
-						if m, ok := el.Obj.(*value.ObjMap); ok {
-							if f, ok := m.Data["fd"]; ok {
-								fd = f.AsInt
-							}
-						} else if inst, ok := el.Obj.(*value.ObjInstance); ok {
-							if f, ok := inst.Fields["fd"]; ok {
-								fd = f.AsInt
-							}
-						}
-
-						if fd != -1 {
-							isReady := false
-							id := int(fd)
-
-							// 1. Check buffers first
-							if _, ok := vm.netBufferedConns[id]; ok {
-								isReady = true
-							} else if _, ok := vm.netBufferedData[id]; ok {
-								isReady = true
-							} else {
-								// 2. Poll
-								vm.shared.NetLock.Lock()
-								l, isListener := vm.shared.NetListeners[id]
-								c, isConn := vm.shared.NetConns[id]
-								vm.shared.NetLock.Unlock()
-
-								if isListener {
-									if tcpL, ok := l.(*net.TCPListener); ok {
-										tcpL.SetDeadline(time.Now().Add(time.Millisecond * time.Duration(timeoutMs)))
-										conn, err := l.Accept()
-										if err == nil {
-											isReady = true
-											vm.netBufferedConns[id] = conn
-										}
-										// Reset deadline?
-										tcpL.SetDeadline(time.Time{})
-									}
-								} else if isConn {
-									conn := c
-									conn.SetReadDeadline(time.Now().Add(time.Millisecond * time.Duration(timeoutMs)))
-									buf := make([]byte, 1) // Peek 1 byte
-									n, err := conn.Read(buf)
-									if err == nil && n > 0 {
-										isReady = true
-										// Buffer the data
-										vm.netBufferedData[id] = buf[:n]
-									}
-									// Reset deadline
-									conn.SetReadDeadline(time.Time{})
-								}
-							}
-
-							if isReady {
-								readyRead = append(readyRead, el)
-							}
-						}
-					}
-				}
-			}
-		}
-
-		// Construct SelectResult Map
-		// struct SelectResult { read: Socket[64], read_count: int, ... }
-
-		// Fill read array up to 64
-		resReadArr := make([]value.Value, 64)
-		for i := 0; i < 64; i++ {
-			if i < len(readyRead) {
-				resReadArr[i] = readyRead[i]
-			} else {
-				resReadArr[i] = value.NewNull()
-			}
-		}
-
-		// Empties for others
-		emptyArr := make([]value.Value, 64)
-		for i := 0; i < 64; i++ {
-			emptyArr[i] = value.NewNull()
-		}
-
-		resFields := map[string]value.Value{
-			"read":        value.NewArray(resReadArr),
-			"read_count":  value.NewInt(int64(len(readyRead))),
-			"write":       value.NewArray(emptyArr),
-			"write_count": value.NewInt(0),
-			"error":       value.NewArray(emptyArr),
-			"error_count": value.NewInt(0),
-		}
-		return value.NewMapWithData(resFields)
-	})
-
-	// SQLite Native Functions
-	vm.DefineNative("sqlite_open", func(args []value.Value) value.Value {
-		if len(args) != 2 {
-			return value.NewNull()
-		} // path, wrapper struct
-		path := args[0].String()
-		structInst, ok := args[1].Obj.(*value.ObjInstance)
-		if !ok {
-			return value.NewNull()
-		}
-		structDef := structInst.Struct
-
-		db, err := sql.Open("sqlite", path)
-		openVal := true
-		if err != nil {
-			openVal = false
-		} else {
-			if err = db.Ping(); err != nil {
-				openVal = false
-			}
-		}
-
-		vm.shared.DbLock.Lock()
-		id := vm.shared.NextDbID
-		vm.shared.NextDbID++
-		vm.shared.DbHandles[id] = db
-		vm.shared.DbLock.Unlock()
-
-		inst := value.NewInstance(structDef).Obj.(*value.ObjInstance)
-		inst.Fields["handle"] = value.NewInt(int64(id))
-		inst.Fields["open"] = value.NewBool(openVal)
-
-		return value.Value{Type: value.VAL_OBJ, Obj: inst}
-	})
-
-	vm.DefineNative("sqlite_close", func(args []value.Value) value.Value {
-		if len(args) != 1 {
-			return value.NewNull()
-		}
-		dbInst, ok := args[0].Obj.(*value.ObjInstance)
-		if !ok {
-			return value.NewNull()
-		}
-
-		handle := int(dbInst.Fields["handle"].AsInt)
-
-		vm.shared.DbLock.Lock()
-		defer vm.shared.DbLock.Unlock()
-
-		if db, ok := vm.shared.DbHandles[handle]; ok {
-			db.Close()
-			delete(vm.shared.DbHandles, handle)
-			dbInst.Fields["open"] = value.NewBool(false)
-		}
-		return value.NewNull()
-	})
-
-	vm.DefineNative("sqlite_exec", func(args []value.Value) value.Value {
-		if len(args) < 3 {
-			return value.NewNull()
-		}
-		dbInst, ok := args[0].Obj.(*value.ObjInstance)
-		if !ok {
-			return value.NewNull()
-		}
-		sqlStr := args[1].String()
-
-		resTmplInst, ok := args[2].Obj.(*value.ObjInstance)
-		if !ok {
-			return value.NewNull()
-		}
-		resStruct := resTmplInst.Struct
-
-		handle := int(dbInst.Fields["handle"].AsInt)
-
-		vm.shared.DbLock.Lock()
-		db, ok := vm.shared.DbHandles[handle]
-		vm.shared.DbLock.Unlock() // Unlock for Exec
-
-		if ok {
-			result, err := db.Exec(sqlStr)
-			resInst := value.NewInstance(resStruct).Obj.(*value.ObjInstance)
-			if err != nil {
-				resInst.Fields["ok"] = value.NewBool(false)
-				resInst.Fields["error"] = value.NewString(err.Error())
-				resInst.Fields["rows_affected"] = value.NewInt(0)
-				resInst.Fields["last_insert_id"] = value.NewInt(0)
-			} else {
-				rowsAffected, _ := result.RowsAffected()
-				lastId, _ := result.LastInsertId()
-				resInst.Fields["ok"] = value.NewBool(true)
-				resInst.Fields["error"] = value.NewString("")
-				resInst.Fields["rows_affected"] = value.NewInt(rowsAffected)
-				resInst.Fields["last_insert_id"] = value.NewInt(lastId)
-			}
-			return value.Value{Type: value.VAL_OBJ, Obj: resInst}
-		}
-		// Invalid handle
-		resInst := value.NewInstance(resStruct).Obj.(*value.ObjInstance)
-		resInst.Fields["ok"] = value.NewBool(false)
-		resInst.Fields["error"] = value.NewString("invalid database handle")
-		resInst.Fields["rows_affected"] = value.NewInt(0)
-		resInst.Fields["last_insert_id"] = value.NewInt(0)
-		return value.Value{Type: value.VAL_OBJ, Obj: resInst}
-	})
-
-	vm.DefineNative("sqlite_exec_params", func(args []value.Value) value.Value {
-		if len(args) < 4 {
-			return value.NewNull()
-		}
-		dbInst, ok := args[0].Obj.(*value.ObjInstance)
-		if !ok {
-			return value.NewNull()
-		}
-		sqlStr := args[1].String()
-		paramsArray, ok := args[2].Obj.(*value.ObjArray)
-		if !ok {
-			return value.NewNull()
-		}
-
-		resTmplInst, ok := args[3].Obj.(*value.ObjInstance)
-		if !ok {
-			return value.NewNull()
-		}
-		resStruct := resTmplInst.Struct
-
-		handle := int(dbInst.Fields["handle"].AsInt)
-
-		vm.shared.DbLock.Lock()
-		db, ok := vm.shared.DbHandles[handle]
-		vm.shared.DbLock.Unlock()
-
-		if ok {
-			// Convert params
-			queryArgs := make([]interface{}, len(paramsArray.Elements))
-			for i, val := range paramsArray.Elements {
-				switch val.Type {
-				case value.VAL_INT:
-					queryArgs[i] = val.AsInt
-				case value.VAL_FLOAT:
-					queryArgs[i] = val.AsFloat
-				case value.VAL_BOOL:
-					queryArgs[i] = val.AsBool
-				case value.VAL_NULL:
-					queryArgs[i] = nil
-				case value.VAL_OBJ:
-					if b, ok := val.Obj.(string); ok {
-						queryArgs[i] = b
-					} else {
-						queryArgs[i] = val.String()
-					}
-				default:
-					queryArgs[i] = val.String()
-				}
-			}
-
-			result, err := db.Exec(sqlStr, queryArgs...)
-			resInst := value.NewInstance(resStruct).Obj.(*value.ObjInstance)
-			if err != nil {
-				resInst.Fields["ok"] = value.NewBool(false)
-				resInst.Fields["error"] = value.NewString(err.Error())
-				resInst.Fields["rows_affected"] = value.NewInt(0)
-				resInst.Fields["last_insert_id"] = value.NewInt(0)
-			} else {
-				rowsAffected, _ := result.RowsAffected()
-				lastId, _ := result.LastInsertId()
-				resInst.Fields["ok"] = value.NewBool(true)
-				resInst.Fields["error"] = value.NewString("")
-				resInst.Fields["rows_affected"] = value.NewInt(rowsAffected)
-				resInst.Fields["last_insert_id"] = value.NewInt(lastId)
-			}
-			return value.Value{Type: value.VAL_OBJ, Obj: resInst}
-		}
-		// Invalid handle
-		resInst := value.NewInstance(resStruct).Obj.(*value.ObjInstance)
-		resInst.Fields["ok"] = value.NewBool(false)
-		resInst.Fields["error"] = value.NewString("invalid database handle")
-		resInst.Fields["rows_affected"] = value.NewInt(0)
-		resInst.Fields["last_insert_id"] = value.NewInt(0)
-		return value.Value{Type: value.VAL_OBJ, Obj: resInst}
-	})
-
-	vm.DefineNative("sqlite_prepare", func(args []value.Value) value.Value {
-		if len(args) < 3 {
-			return value.NewNull()
-		} // db, sql, stmt wrapper
-		dbInst, ok := args[0].Obj.(*value.ObjInstance)
-		if !ok {
-			return value.NewNull()
-		}
-		sqlStr := args[1].String()
-		stmtInst, ok := args[2].Obj.(*value.ObjInstance)
-		if !ok {
-			return value.NewNull()
-		}
-		stmtStructDef := stmtInst.Struct
-
-		handle := int(dbInst.Fields["handle"].AsInt)
-
-		vm.shared.DbLock.Lock()
-		db, ok := vm.shared.DbHandles[handle]
-		vm.shared.DbLock.Unlock()
-
-		if ok {
-			stmt, err := db.Prepare(sqlStr)
-			if err == nil {
-				vm.shared.DbLock.Lock()
-				id := vm.shared.NextStmtID
-				vm.shared.NextStmtID++
-				vm.shared.StmtHandles[id] = stmt
-				vm.shared.StmtParams[id] = make(map[int]interface{})
-				vm.shared.DbLock.Unlock()
-
-				inst := value.NewInstance(stmtStructDef).Obj.(*value.ObjInstance)
-				inst.Fields["handle"] = value.NewInt(int64(id))
-				return value.Value{Type: value.VAL_OBJ, Obj: inst}
-			}
-		}
-		return value.NewNull()
-	})
-
-	bindFunc := func(args []value.Value, val interface{}) value.Value {
-		if len(args) < 3 {
-			return value.NewNull()
-		}
-		stmtInst, ok := args[0].Obj.(*value.ObjInstance)
-		if !ok {
-			return value.NewNull()
-		}
-		idx := int(args[1].AsInt)
-
-		handle := int(stmtInst.Fields["handle"].AsInt)
-
-		vm.shared.DbLock.Lock()
-		defer vm.shared.DbLock.Unlock()
-
-		if _, ok := vm.shared.StmtHandles[handle]; ok {
-			if vm.shared.StmtParams[handle] == nil {
-				vm.shared.StmtParams[handle] = make(map[int]interface{})
-			}
-			vm.shared.StmtParams[handle][idx] = val
-		}
-		return value.NewNull()
-	}
-
-	vm.DefineNative("sqlite_bind_text", func(args []value.Value) value.Value {
-		return bindFunc(args, args[2].String())
-	})
-	vm.DefineNative("sqlite_bind_float", func(args []value.Value) value.Value {
-		return bindFunc(args, args[2].AsFloat)
-	})
-	vm.DefineNative("sqlite_bind_int", func(args []value.Value) value.Value {
-		return bindFunc(args, args[2].AsInt)
-	})
-
-	vm.DefineNative("sqlite_step_exec", func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewNull()
-		}
-		stmtInst, ok := args[0].Obj.(*value.ObjInstance)
-		if !ok {
-			return value.NewNull()
-		}
-		resTmplInst, ok := args[1].Obj.(*value.ObjInstance)
-		if !ok {
-			return value.NewNull()
-		}
-		resStruct := resTmplInst.Struct
-
-		handle := int(stmtInst.Fields["handle"].AsInt)
-
-		vm.shared.DbLock.Lock()
-		stmt, ok := vm.shared.StmtHandles[handle]
-		var params map[int]interface{}
-		if ok {
-			origParams := vm.shared.StmtParams[handle]
-			params = make(map[int]interface{})
-			for k, v := range origParams {
-				params[k] = v
-			}
-		}
-		vm.shared.DbLock.Unlock()
-
-		if ok {
-			// params := vm.stmtParams[handle] // Replaced by copy
-			var maxIdx int
-			for k := range params {
-				if k > maxIdx {
-					maxIdx = k
-				}
-			}
-			argsList := make([]interface{}, maxIdx)
-			for k, v := range params {
-				if k > 0 && k <= maxIdx {
-					argsList[k-1] = v
-				}
-			}
-			result, err := stmt.Exec(argsList...)
-
-			resInst := value.NewInstance(resStruct).Obj.(*value.ObjInstance)
-			if err != nil {
-				resInst.Fields["ok"] = value.NewBool(false)
-				resInst.Fields["error"] = value.NewString(err.Error())
-				resInst.Fields["rows_affected"] = value.NewInt(0)
-				resInst.Fields["last_insert_id"] = value.NewInt(0)
-			} else {
-				rowsAffected, _ := result.RowsAffected()
-				lastId, _ := result.LastInsertId()
-				resInst.Fields["ok"] = value.NewBool(true)
-				resInst.Fields["error"] = value.NewString("")
-				resInst.Fields["rows_affected"] = value.NewInt(rowsAffected)
-				resInst.Fields["last_insert_id"] = value.NewInt(lastId)
-			}
-			return value.Value{Type: value.VAL_OBJ, Obj: resInst}
-		}
-
-		resInst := value.NewInstance(resStruct).Obj.(*value.ObjInstance)
-		resInst.Fields["ok"] = value.NewBool(false)
-		resInst.Fields["error"] = value.NewString("invalid statement handle")
-		resInst.Fields["rows_affected"] = value.NewInt(0)
-		resInst.Fields["last_insert_id"] = value.NewInt(0)
-		return value.Value{Type: value.VAL_OBJ, Obj: resInst}
-	})
-
-	vm.DefineNative("sqlite_reset", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewNull()
-		}
-		stmtInst, ok := args[0].Obj.(*value.ObjInstance)
-		if !ok {
-			return value.NewNull()
-		}
-		handle := int(stmtInst.Fields["handle"].AsInt)
-
-		vm.shared.DbLock.Lock()
-		defer vm.shared.DbLock.Unlock()
-
-		if _, ok := vm.shared.StmtHandles[handle]; ok {
-			vm.shared.StmtParams[handle] = make(map[int]interface{})
-		}
-		return value.NewNull()
-	})
-
-	vm.DefineNative("sqlite_finalize", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewNull()
-		}
-		stmtInst, ok := args[0].Obj.(*value.ObjInstance)
-		if !ok {
-			return value.NewNull()
-		}
-		handle := int(stmtInst.Fields["handle"].AsInt)
-
-		vm.shared.DbLock.Lock()
-		defer vm.shared.DbLock.Unlock()
-
-		if stmt, ok := vm.shared.StmtHandles[handle]; ok {
-			stmt.Close()
-			delete(vm.shared.StmtHandles, handle)
-			delete(vm.shared.StmtParams, handle)
-		}
-		return value.NewNull()
-	})
-
-	vm.DefineNative("sqlite_query", func(args []value.Value) value.Value {
-		if len(args) < 4 {
-			return value.NewNull()
-		} // db, sql, tmplQueryResult, tmplRow
-
-		dbInst, ok := args[0].Obj.(*value.ObjInstance)
-		if !ok {
-			return value.NewNull()
-		}
-		sqlStr := args[1].String()
-
-		resTmplInst, ok := args[2].Obj.(*value.ObjInstance)
-		if !ok {
-			return value.NewNull()
-		}
-		resStruct := resTmplInst.Struct
-
-		rowTmplInst, ok := args[3].Obj.(*value.ObjInstance)
-		if !ok {
-			return value.NewNull()
-		}
-		rowStruct := rowTmplInst.Struct
-
-		handle := int(dbInst.Fields["handle"].AsInt)
-
-		vm.shared.DbLock.Lock()
-		db, ok := vm.shared.DbHandles[handle]
-		vm.shared.DbLock.Unlock()
-
-		if ok {
-			rows, err := db.Query(sqlStr)
-			if err != nil {
-				// Return QueryResult with ok=false and error message
-				resInst := value.NewInstance(resStruct).Obj.(*value.ObjInstance)
-				resInst.Fields["columns"] = value.NewArray(nil)
-				resInst.Fields["rows"] = value.NewArray(nil)
-				resInst.Fields["row_count"] = value.NewInt(0)
-				resInst.Fields["ok"] = value.NewBool(false)
-				resInst.Fields["error"] = value.NewString(err.Error())
-				return value.Value{Type: value.VAL_OBJ, Obj: resInst}
-			}
-			defer rows.Close()
-
-			cols, _ := rows.Columns()
-			colVals := make([]value.Value, len(cols))
-			for i, c := range cols {
-				colVals[i] = value.NewString(c)
-			}
-
-			var rowInsts []value.Value
-
-			for rows.Next() {
-				// Scan to interface{}
-				dest := make([]interface{}, len(cols))
-				destPtrs := make([]interface{}, len(cols))
-				for i := range dest {
-					destPtrs[i] = &dest[i]
-				}
-
-				rows.Scan(destPtrs...)
-
-				rowVals := make([]value.Value, len(cols))
-				for i, v := range dest {
-					// Convert Go type to Noxy value
-					switch tv := v.(type) {
-					case nil:
-						rowVals[i] = value.NewNull()
-					case int64:
-						rowVals[i] = value.NewInt(tv)
-					case float64:
-						rowVals[i] = value.NewFloat(tv)
-					case string:
-						rowVals[i] = value.NewString(tv)
-					case []byte:
-						rowVals[i] = value.NewString(string(tv))
-					default:
-						rowVals[i] = value.NewString(fmt.Sprintf("%v", tv))
-					}
-				}
-
-				// Create Row instance
-				rowInst := value.NewInstance(rowStruct).Obj.(*value.ObjInstance)
-				rowInst.Fields["values"] = value.NewArray(rowVals)
-				rowInsts = append(rowInsts, value.Value{Type: value.VAL_OBJ, Obj: rowInst})
-			}
-
-			// Create QueryResult instance with ok=true
-			resInst := value.NewInstance(resStruct).Obj.(*value.ObjInstance)
-			resInst.Fields["columns"] = value.NewArray(colVals)
-			resInst.Fields["rows"] = value.NewArray(rowInsts)
-			resInst.Fields["row_count"] = value.NewInt(int64(len(rowInsts)))
-			resInst.Fields["ok"] = value.NewBool(true)
-			resInst.Fields["error"] = value.NewString("")
-
-			return value.Value{Type: value.VAL_OBJ, Obj: resInst}
-		}
-		// DB handle not found - return error result
-		resInst := value.NewInstance(resStruct).Obj.(*value.ObjInstance)
-		resInst.Fields["columns"] = value.NewArray(nil)
-		resInst.Fields["rows"] = value.NewArray(nil)
-		resInst.Fields["row_count"] = value.NewInt(0)
-		resInst.Fields["ok"] = value.NewBool(false)
-		resInst.Fields["error"] = value.NewString("invalid database handle")
-		return value.Value{Type: value.VAL_OBJ, Obj: resInst}
-	})
-
-	vm.DefineNative("hex", func(args []value.Value) value.Value {
-		if len(args) != 1 {
-			return value.NewNull()
-		}
-		if args[0].Type == value.VAL_INT {
-			return value.NewString(fmt.Sprintf("0x%x", args[0].AsInt))
-		}
-		if args[0].Type == value.VAL_BYTES {
-			return value.NewString(fmt.Sprintf("%x", args[0].Obj.(string)))
-		}
-		return value.NewString(args[0].String())
-	})
-
-	vm.DefineNative("hex_encode", func(args []value.Value) value.Value {
-		if len(args) != 1 {
-			return value.NewString("")
-		}
-		arg := args[0]
-		var data string
-		if arg.Type == value.VAL_BYTES {
-			data = arg.Obj.(string)
-		} else {
-			data = arg.String()
-		}
-		return value.NewString(hex.EncodeToString([]byte(data)))
-	})
-
-	vm.DefineNative("hex_decode", func(args []value.Value) value.Value {
-		if len(args) != 1 {
-			return value.NewBytes("")
-		}
-		decoded, err := hex.DecodeString(args[0].String())
-		if err != nil {
-			return value.NewBytes("") // Or null/error? Returning empty bytes for simplicity
-		}
-		return value.NewBytes(string(decoded))
-	})
-
-	vm.DefineNative("base64_encode", func(args []value.Value) value.Value {
-		if len(args) != 1 {
-			return value.NewString("")
-		}
-		arg := args[0]
-		var data string
-		if arg.Type == value.VAL_BYTES {
-			data = arg.Obj.(string)
-		} else {
-			data = arg.String()
-		}
-		return value.NewString(base64.StdEncoding.EncodeToString([]byte(data)))
-	})
-
-	vm.DefineNative("base64_decode", func(args []value.Value) value.Value {
-		if len(args) != 1 {
-			return value.NewBytes("")
-		}
-		decoded, err := base64.StdEncoding.DecodeString(args[0].String())
-		if err != nil {
-			return value.NewBytes("")
-		}
-		return value.NewBytes(string(decoded))
-	})
-
-	const base62Chars = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz"
-
-	vm.DefineNative("base62_encode", func(args []value.Value) value.Value {
-		if len(args) != 1 {
-			return value.NewString("")
-		}
-		if args[0].Type != value.VAL_INT {
-			return value.NewString("")
-		}
-		num := args[0].AsInt
-		if num == 0 {
-			return value.NewString("0")
-		}
-
-		var coded []byte
-		neg := false
-		if num < 0 {
-			neg = true
-			num = -num
-		}
-
-		for num > 0 {
-			rem := num % 62
-			coded = append(coded, base62Chars[rem])
-			num = num / 62
-		}
-
-		if neg {
-			coded = append(coded, '-')
-		}
-
-		// Reverse
-		for i, j := 0, len(coded)-1; i < j; i, j = i+1, j-1 {
-			coded[i], coded[j] = coded[j], coded[i]
-		}
-
-		return value.NewString(string(coded))
-	})
-
-	vm.DefineNative("base62_decode", func(args []value.Value) value.Value {
-		if len(args) != 1 {
-			return value.NewInt(0)
-		}
-		str := args[0].String()
-		if str == "" {
-			return value.NewInt(0)
-		}
-		var num int64
-		var neg bool
-		if strings.HasPrefix(str, "-") {
-			neg = true
-			str = str[1:]
-		}
-
-		for _, char := range str {
-			idx := strings.IndexRune(base62Chars, char)
-			if idx == -1 {
-				return value.NewInt(0) // Error
-			}
-			num = num*62 + int64(idx)
-		}
-		if neg {
-			num = -num
-		}
-		return value.NewInt(num)
-	})
-
-	// Precompile regex for fmt verbs
-	// Matches % [flags] [width] [.prec] verb
-	// Flags: [-+ #0]
-	// Width: \d+ or *
-	// Prec: \. followed by \d+ or *
-	// Verb: [a-zA-Z%]
-	fmtVerbRe := regexp.MustCompile(`%([-+ #0]*)(?:(\d+|\*)?)(?:\.(\d+|\*))?([a-zA-Z%])`)
-
-	vm.DefineNative("fmt", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewString("")
-		}
-		formatStr := args[0].String()
-
-		// Parse fmt string to handle %T specifically
-		// We need to rebuild args and format string
-
-		// Find all verbs
-		matches := fmtVerbRe.FindAllStringSubmatchIndex(formatStr, -1)
-
-		var newArgs []interface{}
-		var newFormatBuilder strings.Builder
-
-		argIdx := 0 // Index into args[1:]
-		lastPos := 0
-
-		argsData := args[1:]
-
-		for _, match := range matches {
-			// match indices: [start, end, f_start, f_end, w_start, w_end, p_start, p_end, v_start, v_end]
-			start, end := match[0], match[1]
-			// verb := formatStr[match[8]:match[9]]
-			// We can get verb char easily
-			verb := formatStr[match[8]] // byte
-
-			// Append text before match
-			newFormatBuilder.WriteString(formatStr[lastPos:start])
-
-			// Determine if we need to consume args
-			if verb == '%' {
-				newFormatBuilder.WriteString("%%")
-				lastPos = end
-				continue
-			}
-
-			// Check if width uses arg
-			// Groups are 1-indexed in concept, but indices array:
-			// 0,1 whole
-			// 2,3 flags (group 1)
-			// 4,5 width (group 2)
-			// 6,7 prec (group 3)
-			// 8,9 verb (group 4)
-
-			widthHasStar := false
-			if match[4] != -1 {
-				wStr := formatStr[match[4]:match[5]]
-				if strings.Contains(wStr, "*") {
-					widthHasStar = true
-				}
-			}
-
-			precHasStar := false
-			if match[6] != -1 {
-				pStr := formatStr[match[6]:match[7]]
-				if strings.Contains(pStr, "*") {
-					precHasStar = true
-				}
-			}
-
-			// Consume args for width/prec
-			if widthHasStar {
-				if argIdx < len(argsData) {
-					newArgs = append(newArgs, argsData[argIdx].AsInt) // Assume int for width
-					argIdx++
-				}
-			}
-			if precHasStar {
-				if argIdx < len(argsData) {
-					newArgs = append(newArgs, argsData[argIdx].AsInt) // Assume int for prec
-					argIdx++
-				}
-			}
-
-			// Now handle the verb and the main arg
-			if argIdx < len(argsData) {
-				val := argsData[argIdx]
-				argIdx++
-
-				if verb == 'T' {
-					// Replace %T with %s and supply type name string
-					newFormatBuilder.WriteString("%s")
-
-					typeName := "unknown"
-					switch val.Type {
-					case value.VAL_INT:
-						typeName = "int"
-					case value.VAL_FLOAT:
-						typeName = "float"
-					case value.VAL_BOOL:
-						typeName = "bool"
-					case value.VAL_NULL:
-						typeName = "null"
-					case value.VAL_BYTES:
-						typeName = "bytes"
-					case value.VAL_FUNCTION:
-						typeName = "function"
-					case value.VAL_NATIVE:
-						typeName = "function"
-					case value.VAL_OBJ:
-						// Determine specific object type
-						if _, ok := val.Obj.(*value.ObjArray); ok {
-							typeName = "array"
-						} else if _, ok := val.Obj.(*value.ObjMap); ok {
-							typeName = "map"
-						} else if inst, ok := val.Obj.(*value.ObjInstance); ok {
-							typeName = inst.Struct.Name
-						} else if _, ok := val.Obj.(*value.ObjStruct); ok {
-							typeName = "struct" // Class definition
-						} else if _, ok := val.Obj.(string); ok {
-							typeName = "string"
-						} else {
-							typeName = fmt.Sprintf("%T", val.Obj)
-						}
-					}
-					newArgs = append(newArgs, typeName)
-				} else {
-					// Keep original verb sequence (including flags/width/prec)
-					newFormatBuilder.WriteString(formatStr[start:end])
-
-					// Add argument, potentially wrapped or raw
-					switch val.Type {
-					case value.VAL_INT:
-						newArgs = append(newArgs, val.AsInt)
-					case value.VAL_FLOAT:
-						newArgs = append(newArgs, val.AsFloat)
-					case value.VAL_BOOL:
-						newArgs = append(newArgs, val.AsBool)
-					case value.VAL_NULL:
-						newArgs = append(newArgs, nil)
-					case value.VAL_OBJ:
-						// Pass raw object
-						newArgs = append(newArgs, val.Obj)
-					case value.VAL_BYTES:
-						newArgs = append(newArgs, value.BytesWrapper{Str: val.Obj.(string)})
-					default:
-						newArgs = append(newArgs, val.String())
-					}
-				}
-			} else {
-				// Not enough args? Just append remainder as literal?
-				// Or Go fmt will print %!(MISSING)
-				newFormatBuilder.WriteString(formatStr[start:end])
-			}
-
-			lastPos = end
-		}
-
-		// Append remaining format
-		newFormatBuilder.WriteString(formatStr[lastPos:])
-
-		return value.NewString(fmt.Sprintf(newFormatBuilder.String(), newArgs...))
-	})
-
-	vm.DefineNative("json_dumps", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewString("null")
-		}
-		goVal := jsonValToGo(args[0])
-		bytes, err := json.Marshal(goVal)
-		if err != nil {
-			return value.NewString("null") // Or error?
-		}
-		return value.NewString(string(bytes))
-	})
-
-	// json_parse(str) -> Value
-	vm.DefineNative("json_parse", func(args []value.Value) value.Value {
-		if len(args) < 1 {
-			return value.NewNull()
-		}
-		jsonStr := args[0].String()
-
-		var result interface{}
-		err := json.Unmarshal([]byte(jsonStr), &result)
-		if err != nil {
-			return value.NewNull()
-		}
-		return goValToNoxy(result)
-	})
-
-	// json_loads(str, target) -> Bool
-	jsonLoadsSignature := value.NativeSignature{
-		Arity: 2,
-		Params: []value.ParamInfo{
-			{IsRef: false, TypeName: "string"},
-			{IsRef: true, TypeName: "ref any"},
-		},
-		ReturnType: "bool",
-	}
-	vm.DefineNativeWithSignature("json_loads", jsonLoadsSignature, func(args []value.Value) value.Value {
-		if len(args) < 2 {
-			return value.NewBool(false)
-		}
-		jsonStr := args[0].String()
-		target := args[1]
-
-		var result interface{}
-		decoder := json.NewDecoder(strings.NewReader(jsonStr))
-		decoder.UseNumber()
-		if err := decoder.Decode(&result); err != nil {
-			return value.NewBool(false)
-		}
-		var trailing interface{}
-		if err := decoder.Decode(&trailing); err != io.EOF {
-			return value.NewBool(false)
-		}
-
-		// Try to populate target
-		if populateTarget(vm, target, result) {
-			return value.NewBool(true)
-		}
-		return value.NewBool(false)
-	})
-
+	vm.defineBuiltins()
 	return vm
-}
-
-// Helper: Convert Noxy Value to Go Interface for JSON Marshal
-func jsonValToGo(v value.Value) interface{} {
-	switch v.Type {
-	case value.VAL_NULL:
-		return nil
-	case value.VAL_BOOL:
-		return v.AsBool
-	case value.VAL_INT:
-		return v.AsInt
-	case value.VAL_FLOAT:
-		return v.AsFloat
-	case value.VAL_OBJ:
-		switch o := v.Obj.(type) {
-		case string:
-			return o
-		case *value.ObjArray:
-			arr := make([]interface{}, len(o.Elements))
-			for i, el := range o.Elements {
-				arr[i] = jsonValToGo(el)
-			}
-			return arr
-		case *value.ObjMap:
-			m := make(map[string]interface{})
-			for k, val := range o.Data {
-				keyStr := fmt.Sprintf("%v", k)
-				m[keyStr] = jsonValToGo(val)
-			}
-			return m
-		case *value.ObjInstance:
-			m := make(map[string]interface{})
-			for k, val := range o.Fields {
-				m[k] = jsonValToGo(val)
-			}
-			return m
-		case *value.ObjStruct:
-			return o.Name
-		}
-	case value.VAL_BYTES:
-		// Base64 encode bytes? Or generic string?
-		return v.Obj.(string)
-	}
-	return v.String()
-}
-
-// Helper: Convert Go Interface to Noxy Value
-func goValToNoxy(i interface{}) value.Value {
-	if i == nil {
-		return value.NewNull()
-	}
-	switch v := i.(type) {
-	case bool:
-		return value.NewBool(v)
-	case float64:
-		// JSON numbers are float64 by default
-		// Try to see if it's an int
-		if v == float64(int64(v)) {
-			return value.NewInt(int64(v))
-		}
-		return value.NewFloat(v)
-	case json.Number:
-		if parsed, ok := exactJSONInt(v); ok {
-			return value.NewInt(parsed)
-		}
-		if parsed, err := v.Float64(); err == nil {
-			return value.NewFloat(parsed)
-		}
-		return value.NewNull()
-	case string:
-		return value.NewString(v)
-	case []interface{}:
-		arr := make([]value.Value, len(v))
-		for idx, el := range v {
-			arr[idx] = goValToNoxy(el)
-		}
-		return value.NewArray(arr)
-	case map[string]interface{}:
-		m := make(map[string]value.Value)
-		for k, val := range v {
-			m[k] = goValToNoxy(val)
-		}
-		return value.NewMapWithData(m)
-	}
-	return value.NewString(fmt.Sprintf("%v", i))
-}
-
-// Helper: Populate Target
-func populateTarget(vm *VM, target value.Value, data interface{}) bool {
-	if target.Type == value.VAL_REF {
-		ref, ok := target.Obj.(*value.ObjRef)
-		if !ok || ref == nil {
-			return false
-		}
-		return populateRef(vm, ref, data)
-	} else if target.Type == value.VAL_OBJ {
-		// Populate Object In-Place
-		return populateObj(vm, target, data)
-	}
-	// Cannot populate primitive value passed by value
-	return false
-}
-
-func populateObj(vm *VM, currentVal value.Value, data interface{}) bool {
-	commit, ok := prepareJSONMutation(vm, currentVal, nil, data, nil)
-	if !ok {
-		return false
-	}
-	commit()
-	return true
-}
-
-// Helper: Populate a Reference with Go Data (Deeply)
-func populateRef(vm *VM, ref *value.ObjRef, data interface{}) bool {
-	currentVal, exists, store, err := vm.referenceStorage(ref)
-	if err != nil || !exists || store == nil {
-		return false
-	}
-	if ref.JSONDynamic.Load() {
-		replacement, ok := dynamicJSONValue(data)
-		if !ok {
-			return false
-		}
-		store(replacement)
-		return true
-	}
-	commit, ok := prepareJSONMutation(vm, currentVal, ref.TargetType.Load(), data, jsonSetter(store))
-	if !ok {
-		return false
-	}
-	commit()
-	return true
-}
-
-func jsonReplacementCompatible(current, replacement value.Value) bool {
-	if current.Type == value.VAL_NULL {
-		return true
-	}
-	if current.Type != replacement.Type {
-		return false
-	}
-	if current.Type != value.VAL_OBJ {
-		return true
-	}
-	switch current.Obj.(type) {
-	case string:
-		_, ok := replacement.Obj.(string)
-		return ok
-	case *value.ObjArray:
-		_, ok := replacement.Obj.(*value.ObjArray)
-		return ok
-	case *value.ObjMap:
-		_, ok := replacement.Obj.(*value.ObjMap)
-		return ok
-	case *value.ObjInstance:
-		_, ok := replacement.Obj.(*value.ObjInstance)
-		return ok
-	default:
-		return false
-	}
 }
 
 func (vm *VM) DefineNative(name string, fn value.NativeFunc) {
